Replace old MultiPodBuilder
