+++ src/frontend/mod.rs
@@ -23,7 +23,7 @@ use crate::middleware::{
 
 mod custom;
 mod error;
-mod multi_pod_tree;
+mod multi_pod;
 mod operation;
 mod pod_request;
 mod serialization;
