+++ src/frontend/multi_pod/mod.rs
@@ -1,476 +1,145 @@
-//! Multi-POD builder for automatic statement packing.
+//! Multi-POD builder for the Merkle statement tree design.
 //!
-//! This module provides [`MultiPodBuilder`], a higher-level alternative to [`MainPodBuilder`]
-//! that automatically handles cases where statements exceed per-POD resource limits by
-//! splitting across multiple PODs.
+//! See `docs/multipod_merkle_statement_tree.md` for the design overview.
 //!
-//! # Problem
+//! This is a parallel implementation alongside `super::multi_pod`. The old
+//! solver remains in place during development; once the new POD circuit
+//! lands, this module replaces `multi_pod` and that module is removed.
 //!
-//! A single POD has resource limits (max statements, max custom predicate batches, etc.).
-//! When a proof requires more resources than a single POD can provide, statements must
-//! be split across multiple PODs with dependencies resolved via cross-POD copying.
+//! The solver is purely symbolic: it consumes a positional [`InputShape`]
+//! and produces a positional [`OutputShape`]. The build layer in this
+//! module translates between user-facing builder state and the symbolic
+//! representation, holding an internal side table that maps positional
+//! indices back to concrete pod hashes and external (input) statements.
 //!
-//! # Architecture
+//! `solve()` partitions the workload symbolically; `prove()` then walks
+//! the partition and builds + proves each POD in chain order.
 //!
-//! The multi-POD system uses a MILP (Mixed Integer Linear Program) solver to find the
-//! optimal assignment of statements to PODs. The solver minimizes the number of PODs
-//! while respecting:
-//! - Per-POD resource limits (statements, batches, merkle proofs, etc.)
-//! - Statement dependencies (if A depends on B, B must be available when proving A)
-//! - Input POD limits (each POD can only reference a limited number of other PODs)
-//!
-//! # POD Ordering
-//!
-//! PODs are built in index order: 0, 1, 2, ..., k. The **output POD is always last**
-//! (index k), containing the user-requested public statements. Earlier PODs (0..k-1)
-//! are **intermediate PODs** that prove supporting statements.
-//!
-//! This ordering allows dependencies to flow forward: later PODs can access public
-//! statements from earlier PODs by adding them as input PODs. The output POD, being
-//! last, can access all intermediate PODs.
-//!
-//! # Usage
-//!
-//! ```ignore
-//! let mut builder = MultiPodBuilder::new(&params, &vd_set);
-//!
-//! // Add operations (similar to MainPodBuilder)
-//! let stmt_a = builder.priv_op(FrontendOp::eq(1, 1))?;
-//! let stmt_b = builder.pub_op(FrontendOp::eq(2, 2))?;  // Will be public in output
-//!
-//! // Solve and prove
-//! let result = builder.prove(&prover)?;
-//!
-//! // Access the output POD
-//! let output = result.output_pod();
-//! ```
-//!
-//! [`MainPodBuilder`]: crate::frontend::MainPodBuilder
+//! The module is not yet re-exported from `frontend::mod`; until it is,
+//! the public surface is reachable only from tests, so we silence
+//! dead-code warnings module-wide.
+
+#![allow(dead_code)]
 
 use std::{
-    collections::{BTreeSet, HashMap},
+    collections::{hash_map::Entry, BTreeSet, HashMap, HashSet},
     fmt,
 };
 
 use crate::{
     frontend::{MainPod, MainPodBuilder, Operation},
-    middleware::{Hash, MainPodProver, Params, Statement, VDSet, Value},
+    middleware::{
+        Hash, InputPodOpenStatement, MainPodProver, NativeOperation, OperationAux, OperationType,
+        Params, Statement, VDSet, Value, BASE_PARAMS,
+    },
 };
 
 mod cost;
 mod deps;
-pub mod diagnostics;
-mod solver;
+mod diagnostics;
+mod partition;
+#[cfg(test)]
+mod partition_milp;
+mod shape;
 
 use cost::StatementCost;
-use deps::{DependencyGraph, StatementSource};
+use deps::{DependencyGraph, ExternalDependency, StatementSource};
 pub use diagnostics::{ResourceSummary, SolutionBreakdown};
-pub use solver::MultiPodSolution;
+pub use shape::{AbstractDep, InputShape, OutputShape};
 
-/// Error type for multi-POD operations.
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     Custom(String),
-    /// Error from the frontend.
     Frontend(#[from] crate::frontend::Error),
-    /// Error from the MILP solver.
-    Solver(String),
-    /// No solution exists (shouldn't happen with valid input).
-    NoSolution,
+    NoFeasiblePartition(String),
+    ChainTreeCapacityExceeded { needed: usize, capacity: usize },
 }
 
-impl std::fmt::Display for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Error::Custom(msg) => write!(f, "Custom error: {}", msg),
             Error::Frontend(e) => write!(f, "Frontend error: {}", e),
-            Error::Solver(msg) => write!(f, "Solver error: {}", msg),
-            Error::NoSolution => write!(f, "No solution exists"),
+            Error::NoFeasiblePartition(msg) => {
+                write!(f, "No feasible partition: {}", msg)
+            }
+            Error::ChainTreeCapacityExceeded { needed, capacity } => write!(
+                f,
+                "Chain tree capacity exceeded: partition would publish {} \
+                 statements across the chain but the public-statements MT \
+                 has capacity {}",
+                needed, capacity
+            ),
         }
     }
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-/// Default maximum number of PODs the solver will consider.
-pub const DEFAULT_MAX_PODS: usize = 20;
-
-/// Options for configuring MultiPodBuilder behavior.
-#[derive(Debug, Clone)]
-pub struct Options {
-    /// Maximum number of PODs the solver will consider.
-    /// Defaults to 20. Increase if you have a very large number of statements.
-    pub max_pods: usize,
+/// Side table pairing an [`OutputShape`]'s positional external indices
+/// with the concrete pod hashes and input statements they refer to. The
+/// solver never sees concrete data; the build layer uses this index to
+/// reattach hashes when materialising a [`MultiPodResult`] from a
+/// partition.
+#[derive(Clone, Debug)]
+struct ExternalIndex {
+    pods: Vec<Hash>,
+    statements: Vec<ExternalDependency>,
+    /// Inverse of `pods` for O(1) hash → abstract-pod-index lookup.
+    pod_index_by_hash: HashMap<Hash, usize>,
 }
 
-impl Default for Options {
-    fn default() -> Self {
+impl ExternalIndex {
+    fn new(pods: Vec<Hash>, statements: Vec<ExternalDependency>) -> Self {
+        let pod_index_by_hash = pods.iter().copied().zip(0..).collect();
         Self {
-            max_pods: DEFAULT_MAX_PODS,
+            pods,
+            statements,
+            pod_index_by_hash,
         }
     }
 }
 
-/// Result of proving with MultiPodBuilder.
-#[derive(Debug)]
-pub struct MultiPodResult {
-    /// All PODs in build order (0, 1, ..., k).
-    /// Intermediate PODs are at indices 0..k-1.
-    /// The output POD is at index k (the last POD).
-    pub pods: Vec<MainPod>,
-}
-
-impl MultiPodResult {
-    /// Get the output POD (containing user-requested public statements).
-    /// This is always the last POD (`pods[k]`), which can access all earlier
-    /// intermediate PODs for dependencies.
-    pub fn output_pod(&self) -> &MainPod {
-        self.pods
-            .last()
-            .expect("MultiPodResult must have at least one POD")
-    }
-
-    /// Get intermediate/supporting PODs (all PODs except the output POD).
-    /// These are at indices 0..k-1, built before the output POD.
-    pub fn intermediate_pods(&self) -> &[MainPod] {
-        &self.pods[..self.pods.len() - 1]
-    }
-}
-
-/// Builder for creating multiple PODs when statements exceed per-POD limits.
-///
-/// # Overview
-///
-/// `MultiPodBuilder` provides a similar API to [`MainPodBuilder`], but automatically
-/// splits statements across multiple PODs when resource limits are exceeded. The
-/// workflow is:
-///
-/// 1. **Add operations**: Use [`priv_op`](Self::priv_op) and [`pub_op`](Self::pub_op)
-///    to add statements, just like `MainPodBuilder`.
-///
-/// 2. **Solve**: Call [`solve`](Self::solve) to run the MILP solver, which determines
-///    the optimal assignment of statements to PODs. This consumes the builder and
-///    returns a [`SolvedMultiPod`].
-///
-/// 3. **Prove**: Call [`prove`](SolvedMultiPod::prove) on the solved result to build
-///    and prove all PODs.
-///
-/// # POD Structure
-///
-/// The result contains PODs in build order: intermediate PODs first (indices 0..k-1),
-/// then the output POD last (index k). The output POD contains all user-requested
-/// public statements (those added via `pub_op`). Intermediate PODs make their
-/// statements public so later PODs can access them.
-///
-/// [`MainPodBuilder`]: crate::frontend::MainPodBuilder
-#[derive(Debug)]
 pub struct MultiPodBuilder {
     params: Params,
     vd_set: VDSet,
-    options: Options,
-    /// External input PODs (already proved).
     input_pods: Vec<MainPod>,
-    /// Optional initial wildcard values for custom operations
     operations_wildcard_values: HashMap<usize, Vec<(usize, Value)>>,
-    /// Indices of statements that should be public in output PODs.
-    /// Uses Vec since max_public_statements is small (≤8); indices are naturally sorted.
     output_public_indices: Vec<usize>,
-    /// Used during add_operation to validate statements with unlimited params.
     builder: MainPodBuilder,
 }
 
-/// A solved multi-POD problem, ready to be proved.
-///
-/// Created by [`MultiPodBuilder::solve`]. Call [`prove`](Self::prove) to build
-/// and prove all PODs, or inspect the [`solution`](Self::solution) first.
-#[derive(Debug)]
-pub struct SolvedMultiPod {
-    params: Params,
-    vd_set: VDSet,
-    input_pods: Vec<MainPod>,
-    statements: Vec<Statement>,
-    operations: Vec<Operation>,
-    output_public_indices: Vec<usize>,
-    operations_wildcard_values: HashMap<usize, Vec<(usize, Value)>>,
-    solution: MultiPodSolution,
-    deps: DependencyGraph,
-}
-
-impl SolvedMultiPod {
-    /// Get the solver's solution (POD assignments).
-    pub fn solution(&self) -> &MultiPodSolution {
-        &self.solution
-    }
-
-    /// Compute a post-solve per-POD resource utilization breakdown.
-    pub fn solution_breakdown(&self) -> SolutionBreakdown {
-        let costs: Vec<StatementCost> = self
-            .operations
-            .iter()
-            .map(StatementCost::from_operation)
-            .collect();
-        SolutionBreakdown::from_solution(
-            &costs,
-            &self.solution.pod_statements,
-            self.solution.pod_count,
-            self.statements.len(),
-            &self.params,
-        )
-    }
-
-    /// Build and prove all PODs.
-    ///
-    /// Builds PODs in dependency order (0, 1, ..., k) and proves each one.
-    /// The last POD is the output POD containing user-requested public statements.
-    pub fn prove(self, prover: &dyn MainPodProver) -> Result<MultiPodResult> {
-        let solution = &self.solution;
-
-        // Build PODs in sequential order: 0, 1, 2, ..., k
-        // This order is guaranteed by the solver's symmetry-breaking constraint.
-        // Later PODs may reference earlier ones for cross-POD dependencies.
-        let mut pods: Vec<MainPod> = Vec::with_capacity(solution.pod_count);
-
-        for pod_idx in 0..solution.pod_count {
-            let pod = self.build_single_pod(pod_idx, &pods, prover)?;
-            pods.push(pod);
-        }
-
-        Ok(MultiPodResult { pods })
-    }
-
-    /// Build a single POD based on the solver solution.
-    ///
-    /// This function translates the solver's abstract assignment into a concrete POD by:
-    /// 1. Identifying which input PODs are needed (external + earlier generated)
-    /// 2. Adding those input PODs to a fresh `MainPodBuilder`
-    /// 3. For each statement assigned to this POD (in dependency order):
-    ///    - Execute the original operation to create the statement
-    ///    - Mark as public if the solver determined it should be
-    /// 4. Prove the POD
-    fn build_single_pod(
-        &self,
-        pod_idx: usize,
-        earlier_pods: &[MainPod],
-        prover: &dyn MainPodProver,
-    ) -> Result<MainPod> {
-        let mut builder = MainPodBuilder::new(&self.params, &self.vd_set);
-        let solution = &self.solution;
-        let statements_in_this_pod = &solution.pod_statements[pod_idx];
-
-        // Step 1: Find which external and earlier PODs we need based on dependencies
-        let (needed_earlier_pods, needed_external_pods) = self.compute_pod_inputs(pod_idx);
-
-        // Step 2: Add input PODs to the builder
-        for ext_idx in needed_external_pods {
-            builder.add_pod(self.input_pods[ext_idx].clone())?;
-        }
-        for earlier_idx in needed_earlier_pods {
-            builder.add_pod(earlier_pods[earlier_idx].clone())?;
-        }
-
-        // Step 3: Add statements in dependency order.
-        // Statements are added in ascending index order, which matches dependency order:
-        // if B depends on A, then A has a lower index and is added first.
-        let statements_sorted: BTreeSet<usize> = statements_in_this_pod.iter().copied().collect();
-        let public_set = &solution.pod_public_statements[pod_idx];
-
-        for &stmt_idx in &statements_sorted {
-            let op = self.operations[stmt_idx].clone();
-            let wildcard_values = self
-                .operations_wildcard_values
-                .get(&stmt_idx)
-                .cloned()
-                .unwrap_or_default();
-
-            let stmt = builder.op(false, wildcard_values, op)?;
-            assert_eq!(stmt, self.statements[stmt_idx]); // Sanity check
-        }
-
-        // For the output pod, make statements public in the original order.
-        // Intermediate pods use the solver-selected public set.
-        if pod_idx == solution.pod_count - 1 {
-            for idx in &self.output_public_indices {
-                builder.reveal(&self.statements[*idx])?;
-            }
-        } else {
-            for idx in public_set {
-                builder.reveal(&self.statements[*idx])?;
-            }
-        }
-
-        // Forward external premises only when the solver selected them as public
-        // for this POD. These do not require local proving in this POD.
-        for ext_premise_idx in &solution.pod_public_external_premises[pod_idx] {
-            let ext_premise = &solution.external_premises[*ext_premise_idx];
-            builder.reveal(&ext_premise.statement)?;
-        }
-
-        // Step 4: Prove the POD
-        let pod = builder.prove(prover)?;
-
-        Ok(pod)
-    }
-
-    /// Compute which input PODs (internal and external) are needed for a given POD.
-    ///
-    /// Returns (internal_pod_indices, external_pod_indices).
-    fn compute_pod_inputs(&self, pod_idx: usize) -> (BTreeSet<usize>, BTreeSet<usize>) {
-        let solution = &self.solution;
-        let internal_pods = solution.pod_internal_inputs[pod_idx].clone();
-        let mut external_pods: BTreeSet<usize> = BTreeSet::new();
-
-        for external_idx in &solution.pod_external_inputs[pod_idx] {
-            let pod_hash = solution.external_pod_hashes[*external_idx];
-            let idx = self
-                .input_pods
-                .iter()
-                .position(|p| p.statements_hash() == pod_hash)
-                .expect("external pod hash from solver solution");
-            external_pods.insert(idx);
-        }
-
-        assert!(internal_pods.len() + external_pods.len() <= self.params.max_input_pods);
-
-        (internal_pods, external_pods)
-    }
-}
-
-impl fmt::Display for SolvedMultiPod {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let solution = &self.solution;
-        let output_pod_idx = solution.pod_count.saturating_sub(1);
-
-        // Header
-        writeln!(
-            f,
-            "SolvedMultiPod: {} statements → {} PODs",
-            self.statements.len(),
-            solution.pod_count
-        )?;
-
-        if !self.input_pods.is_empty() {
-            writeln!(f, "  External input PODs: {}", self.input_pods.len())?;
-        }
-
-        writeln!(f)?;
-
-        // Per-POD breakdown
-        for pod_idx in 0..solution.pod_count {
-            let is_output = pod_idx == output_pod_idx;
-            let role = if is_output { "output" } else { "intermediate" };
-
-            writeln!(f, "  POD {} ({}):", pod_idx, role)?;
-
-            // Show input PODs
-            let (internal_inputs, external_inputs) = self.compute_pod_inputs(pod_idx);
-            if !internal_inputs.is_empty() || !external_inputs.is_empty() {
-                let internal_str: Vec<String> = internal_inputs
-                    .iter()
-                    .map(|i| format!("POD {}", i))
-                    .collect();
-                let external_str: Vec<String> = external_inputs
-                    .iter()
-                    .map(|i| format!("ext[{}]", i))
-                    .collect();
-                let all_inputs: Vec<&str> = internal_str
-                    .iter()
-                    .map(|s| s.as_str())
-                    .chain(external_str.iter().map(|s| s.as_str()))
-                    .collect();
-                writeln!(
-                    f,
-                    "    inputs: {} (total: {})",
-                    all_inputs.join(", "),
-                    all_inputs.len()
-                )?;
-            }
-
-            // Show statements
-            let stmts = &solution.pod_statements[pod_idx];
-            let public_stmts = &solution.pod_public_statements[pod_idx];
-
-            for &stmt_idx in stmts {
-                let stmt = &self.statements[stmt_idx];
-                let is_public = public_stmts.contains(&stmt_idx);
-                let visibility = if is_public { "public" } else { "private" };
-
-                // Show dependencies for this statement
-                let deps = &self.deps.statement_deps[stmt_idx];
-                let dep_str = if deps.is_empty() {
-                    String::new()
-                } else {
-                    let dep_parts: Vec<String> = deps
-                        .iter()
-                        .map(|d| match d {
-                            StatementSource::Internal(i) => format!("stmt[{}]", i),
-                            StatementSource::External(_) => "ext".to_string(),
-                        })
-                        .collect();
-                    format!(" ← {}", dep_parts.join(", "))
-                };
-
-                writeln!(f, "    [{}] {} [{}]{}", stmt_idx, stmt, visibility, dep_str)?;
-            }
-
-            writeln!(f)?;
-        }
-
-        Ok(())
-    }
-}
-
 impl MultiPodBuilder {
-    /// Create a new MultiPodBuilder with default options.
     pub fn new(params: &Params, vd_set: &VDSet) -> Self {
-        Self::new_with_options(params, vd_set, Options::default())
-    }
-
-    /// Create a new MultiPodBuilder with custom options.
-    pub fn new_with_options(params: &Params, vd_set: &VDSet, options: Options) -> Self {
         let unlimited_params = Params {
             max_statements: usize::MAX / 2,
-            max_public_statements: usize::MAX / 2,
             max_input_pods: usize::MAX / 2,
-            max_input_pods_public_statements: usize::MAX / 2,
+            max_open_input_statements: usize::MAX / 2,
             ..params.clone()
         };
         let builder = MainPodBuilder::new(&unlimited_params, vd_set);
         Self {
             params: params.clone(),
             vd_set: vd_set.clone(),
-            options,
-            builder,
             input_pods: Vec::new(),
             operations_wildcard_values: HashMap::new(),
             output_public_indices: Vec::new(),
+            builder,
         }
     }
 
-    /// Add an external input POD.
     pub fn add_pod(&mut self, pod: MainPod) -> Result<()> {
         self.builder.add_pod(pod.clone())?;
         self.input_pods.push(pod);
         Ok(())
     }
 
-    /// Add a public operation (statement will be public in output).
     pub fn pub_op(&mut self, op: Operation) -> Result<Statement> {
         self.op(true, vec![], op)
     }
 
-    /// Add a private operation.
     pub fn priv_op(&mut self, op: Operation) -> Result<Statement> {
         self.op(false, vec![], op)
-    }
-
-    // Find the index of a statement that has been added.  Panics if the statement doesn't
-    // exist.
-    fn stmt_index(&self, stmt: &Statement) -> usize {
-        self.builder
-            .statements
-            .iter()
-            .position(|s| s == stmt)
-            .expect("exists")
     }
 
     pub fn op(
@@ -479,150 +148,189 @@ impl MultiPodBuilder {
         wildcard_values: Vec<(usize, Value)>,
         op: Operation,
     ) -> Result<Statement> {
-        let stmt = self.add_operation(wildcard_values, op)?;
+        let (stmt, idx) = self.add_operation(wildcard_values, op)?;
         if public {
-            let index = self.stmt_index(&stmt);
-            if !self.output_public_indices.contains(&index) {
-                self.output_public_indices.push(index);
-            }
+            self.mark_public(idx);
         }
         Ok(stmt)
     }
 
-    /// Internal: Add an operation and create its statement.
     fn add_operation(
         &mut self,
         wildcard_values: Vec<(usize, Value)>,
         op: Operation,
-    ) -> Result<Statement> {
-        // Get or create the cached builder
-        //
-        // NOTE: We clone input pods here because MainPodBuilder takes ownership.
-        // This could be avoided if MainPodBuilder were generic over the pod storage type:
-        //   struct MainPodBuilder<P: Borrow<MainPod> = MainPod>
-        // Then MultiPodBuilder could use MainPodBuilder<&MainPod> to borrow instead of clone,
-        // while existing code using MainPodBuilder (with the default) would be unaffected.
-        let stmt = self
-            .builder
-            .op(false, wildcard_values.clone(), op.clone())?;
-        self.operations_wildcard_values
-            .insert(self.stmt_index(&stmt), wildcard_values.clone());
-
-        Ok(stmt)
+    ) -> Result<(Statement, usize)> {
+        let stmt = self.builder.op(false, wildcard_values.clone(), op)?;
+        let idx = self.stmt_index(&stmt);
+        self.operations_wildcard_values.insert(idx, wildcard_values);
+        Ok((stmt, idx))
     }
 
-    /// Mark a statement as public in output.
-    ///
-    /// Returns an error if the statement was not found in the builder.
-    /// Calling this multiple times on the same statement is idempotent.
-    pub fn reveal(&mut self, stmt: &Statement) -> Result<()> {
-        if let Some(idx) = self.builder.statements.iter().position(|s| s == stmt) {
-            if !self.output_public_indices.contains(&idx) {
-                self.output_public_indices.push(idx);
-            }
-            Ok(())
-        } else {
-            Err(Error::Custom(
-                "reveal() called with statement not found in builder".to_string(),
-            ))
+    fn stmt_index(&self, stmt: &Statement) -> usize {
+        self.builder
+            .statements
+            .iter()
+            .position(|(_, s)| s == stmt)
+            .expect("statement exists in builder")
+    }
+
+    fn mark_public(&mut self, idx: usize) {
+        if !self.output_public_indices.contains(&idx) {
+            self.output_public_indices.push(idx);
         }
     }
 
-    /// Get the number of statements.
+    pub fn reveal(&mut self, stmt: &Statement) -> Result<()> {
+        let idx = self
+            .builder
+            .statements
+            .iter()
+            .position(|(_, s)| s == stmt)
+            .ok_or_else(|| {
+                Error::Custom("reveal() called with statement not found in builder".to_string())
+            })?;
+        self.mark_public(idx);
+        Ok(())
+    }
+
     pub fn stmt_len(&self) -> usize {
         self.builder.stmt_len()
     }
 
-    /// Skip the MILP partitioner and prove a single MainPod directly
-    /// via the inner `MainPodBuilder`. Useful for fixture-capture flows
-    /// where the multi-POD decomposition is irrelevant and the slow
-    /// solver dominates wall time. The resulting pod will exceed real
-    /// per-POD limits and is only safe under `MockProver` (the internal
-    /// builder is constructed with `usize::MAX / 2` for all per-POD
-    /// caps, so it accepts arbitrary statement counts).
-    pub fn prove_into_single_pod(mut self, prover: &dyn MainPodProver) -> Result<MainPod> {
-        let statements_to_reveal: Vec<Statement> = self
-            .output_public_indices
+    /// Read-only view of the inner builder's recorded operations. Same
+    /// indexing as [`InputShape::costs`] / [`InputShape::dep_edges`], so
+    /// callers can map a positional cost back to the operation that
+    /// produced it.
+    pub fn operations(&self) -> &[Operation] {
+        &self.builder.operations
+    }
+
+    /// Snapshot of the inner builder's recorded statements. Same
+    /// indexing as [`Self::operations`].
+    pub fn statements(&self) -> Vec<Statement> {
+        self.builder
+            .statements
             .iter()
-            .map(|&i| self.builder.statements[i].clone())
+            .map(|(_, s)| s.clone())
+            .collect()
+    }
+
+    /// Read-only view of the input PODs this builder will reference as
+    /// external sources. Indices match the `pod` field of
+    /// [`AbstractDep::External`] entries in [`Self::input_shape`].
+    pub fn input_pods(&self) -> &[MainPod] {
+        &self.input_pods
+    }
+
+    /// Compute the symbolic [`InputShape`] without running the
+    /// partitioner. Useful for capturing fixtures from a real builder
+    /// run and inspecting the shape before solving.
+    pub fn input_shape(&self) -> InputShape {
+        let external_pod_statements = build_external_statement_map(&self.input_pods);
+        let statements: Vec<Statement> = self
+            .builder
+            .statements
+            .iter()
+            .map(|(_, s)| s.clone())
             .collect();
-        for st in statements_to_reveal {
-            self.builder.reveal(&st)?;
-        }
-        self.builder.prove(prover).map_err(Error::from)
+        let deps = DependencyGraph::build(
+            &statements,
+            &self.builder.operations,
+            &external_pod_statements,
+        );
+        let (shape, _) = build_shape_and_index(
+            &self.builder.operations,
+            &deps,
+            &self.output_public_indices,
+            &self.params,
+        );
+        shape
     }
 
-    /// Replay this builder's accumulated operations onto a fresh
-    /// [`multi_pod_tree::MultiPodBuilder`](crate::frontend::multi_pod_tree::MultiPodBuilder).
-    /// Useful for side-by-side debugging and fixture capture: the
-    /// resulting tree builder represents the same problem and can be
-    /// inspected via `input_shape()` without running the partitioner.
-    pub fn replay_into_tree(&self) -> crate::frontend::multi_pod_tree::MultiPodBuilder {
-        let mut tree =
-            crate::frontend::multi_pod_tree::MultiPodBuilder::new(&self.params, &self.vd_set);
-        for pod in &self.input_pods {
-            tree.add_pod(pod.clone()).expect("add_pod replays cleanly");
-        }
-        for (idx, op) in self.builder.operations.iter().enumerate() {
-            let public = self.output_public_indices.contains(&idx);
-            let wildcard_values = self
-                .operations_wildcard_values
-                .get(&idx)
-                .cloned()
-                .unwrap_or_default();
-            tree.op(public, wildcard_values, op.clone())
-                .expect("op replays cleanly");
-        }
-        tree
-    }
-
-    /// Compute a pre-solve resource summary showing aggregate demand vs. per-POD limits.
-    ///
-    /// This is useful for understanding which resource category is the bottleneck
-    /// before running the solver, especially when debugging solver performance issues.
+    /// Pre-solve resource summary. Aggregates per-statement costs against
+    /// per-POD limits and identifies the bottleneck resource.
     pub fn resource_summary(&self) -> ResourceSummary {
         let costs: Vec<StatementCost> = self
             .builder
             .operations
             .iter()
-            .map(StatementCost::from_operation)
+            .map(|op| StatementCost::from_operation(op, &self.params))
             .collect();
-        ResourceSummary::from_costs(&costs, &self.params)
+        ResourceSummary::from_costs(costs.iter(), &self.params)
     }
 
-    /// Solve the packing problem and return a solved builder ready for proving.
-    ///
-    /// This runs the MILP solver to find the optimal POD assignment.
-    /// Consumes the builder and returns a [`SolvedMultiPod`] that can be proved.
+    /// Solve the partitioning problem. Builds the [`InputShape`] from the
+    /// current builder state (including the external-republish pre-pass),
+    /// runs the DP partitioner, and returns a [`SolvedMultiPod`] that
+    /// holds the partition and the side data needed to materialise PODs
+    /// (once the circuit is available).
     pub fn solve(self) -> Result<SolvedMultiPod> {
         let MainPodBuilder {
-            statements,
+            statements: stmt_pairs,
             operations,
             ..
         } = self.builder;
-        // Compute costs for each statement
-        let costs: Vec<StatementCost> = operations
-            .iter()
-            .map(StatementCost::from_operation)
-            .collect();
+        let statements: Vec<Statement> = stmt_pairs.into_iter().map(|(_, s)| s).collect();
 
-        // Build external POD statement mapping
         let external_pod_statements = build_external_statement_map(&self.input_pods);
-
-        // Build dependency graph
         let deps = DependencyGraph::build(&statements, &operations, &external_pod_statements);
 
-        // Run solver
-        let input = solver::SolverInput {
-            num_statements: statements.len(),
-            costs: &costs,
-            deps: &deps,
-            output_public_indices: &self.output_public_indices,
-            params: &self.params,
-            max_pods: self.options.max_pods,
-        };
+        let (shape, external_index) = build_shape_and_index(
+            &operations,
+            &deps,
+            &self.output_public_indices,
+            &self.params,
+        );
 
-        let solution = solver::solve(&input)?;
+        let mut output = partition::partition(&shape).ok_or_else(|| {
+            Error::NoFeasiblePartition(
+                "the DP partitioner could not find a feasible partition under \
+                 the current params; run diagnose_failure() for details"
+                    .to_string(),
+            )
+        })?;
+
+        // Synthetic republish statements appear at positions >= operations.len()
+        // in the augmented shape, each with one External dep recording its
+        // input statement. Whichever POD a synthetic lands in republishes
+        // that input statement.
+        let n_orig = operations.len();
+        let mut new_republished: Vec<BTreeSet<usize>> = vec![BTreeSet::new(); output.pod_count];
+        for (pod_idx, stmts) in output.pod_statements.iter().enumerate() {
+            for &s in stmts {
+                if s >= n_orig {
+                    if let AbstractDep::External { statement, .. } = shape.dep_edges[s][0] {
+                        new_republished[pod_idx].insert(statement);
+                    }
+                }
+            }
+        }
+        output.pod_republished_externals = new_republished;
+
+        let public_sets = intermediate_public_sets(&shape, &output);
+        let chain_total: usize = public_sets
+            .iter()
+            .take(output.pod_count.saturating_sub(1))
+            .map(|s| s.len())
+            .sum();
+        let chain_capacity = 2 << BASE_PARAMS.max_depth_public_statements_mt;
+        if chain_total > chain_capacity {
+            return Err(Error::ChainTreeCapacityExceeded {
+                needed: chain_total,
+                capacity: chain_capacity,
+            });
+        }
+
+        let input_pod_idx_by_abs: Vec<usize> = external_index
+            .pods
+            .iter()
+            .map(|h| {
+                self.input_pods
+                    .iter()
+                    .position(|p| p.statements_hash() == *h)
+                    .expect("external pod referenced by user op is in input_pods")
+            })
+            .collect();
 
         Ok(SolvedMultiPod {
             params: self.params,
@@ -632,13 +340,259 @@ impl MultiPodBuilder {
             operations,
             output_public_indices: self.output_public_indices,
             operations_wildcard_values: self.operations_wildcard_values,
-            solution,
-            deps,
+            shape,
+            output,
+            external_index,
+            input_pod_idx_by_abs,
+            public_sets,
         })
     }
 }
 
-/// Build mapping from external POD statements to their POD hash.
+/// A solved multi-POD problem. Carries the partition plus everything
+/// needed to materialise concrete PODs once the new circuit lands.
+pub struct SolvedMultiPod {
+    params: Params,
+    vd_set: VDSet,
+    input_pods: Vec<MainPod>,
+    statements: Vec<Statement>,
+    operations: Vec<Operation>,
+    output_public_indices: Vec<usize>,
+    operations_wildcard_values: HashMap<usize, Vec<(usize, Value)>>,
+    shape: InputShape,
+    output: OutputShape,
+    external_index: ExternalIndex,
+    /// `external_index.pods[abs_pod]` is a hash; this maps that
+    /// abstract index to the matching POD's position in `input_pods`,
+    /// so `pod_inputs` can attach the right `MainPod` without scanning.
+    input_pod_idx_by_abs: Vec<usize>,
+    /// Per-POD public sets. Computed once at `solve()` for the chain-tree
+    /// capacity check and reused by `prove()`.
+    public_sets: Vec<BTreeSet<usize>>,
+}
+
+impl SolvedMultiPod {
+    pub fn input_shape(&self) -> &InputShape {
+        &self.shape
+    }
+
+    pub fn solution(&self) -> &OutputShape {
+        &self.output
+    }
+
+    pub fn solution_breakdown(&self) -> SolutionBreakdown {
+        SolutionBreakdown::from_solution(&self.shape, &self.output)
+    }
+
+    /// Number of original user-added operations. Statement positions in
+    /// `solution().pod_statements` at or beyond this index are synthetic
+    /// republish statements (zero cost, one external dep).
+    pub fn num_original_statements(&self) -> usize {
+        self.operations.len()
+    }
+
+    /// Build and prove all PODs in the partition in chain order. Each
+    /// intermediate POD extends its chain predecessor's public statement
+    /// tree (slot 0) and appends only the statements downstream consumers
+    /// will open. The output POD does not extend; it materialises a fresh
+    /// public statement tree containing exactly `output_public_indices`,
+    /// importing any upstream-produced output-publics through slot 0.
+    pub fn prove(self, prover: &dyn MainPodProver) -> Result<MultiPodResult> {
+        let pod_count = self.output.pod_count;
+        let mut pods: Vec<MainPod> = Vec::with_capacity(pod_count);
+        for p in 0..pod_count {
+            let pod = self.build_single_pod(p, &pods, prover)?;
+            pods.push(pod);
+        }
+        Ok(MultiPodResult { pods })
+    }
+
+    fn build_single_pod(
+        &self,
+        p: usize,
+        earlier_pods: &[MainPod],
+        prover: &dyn MainPodProver,
+    ) -> Result<MainPod> {
+        let is_output = p + 1 == self.output.pod_count;
+        let mut builder = MainPodBuilder::new(&self.params, &self.vd_set);
+
+        let (inputs, ext_slot) = self.pod_inputs(p, earlier_pods);
+        for input in inputs {
+            builder.add_pod(input)?;
+        }
+
+        if p > 0 && !is_output {
+            builder.extend_input_pod0_public_statements();
+        }
+
+        let n_orig = self.operations.len();
+        let public_set = &self.public_sets[p];
+        for &s in &self.output.pod_statements[p] {
+            if s >= n_orig {
+                // Synthetic republish: open the input statement so
+                // downstream PODs can read it from the chain at slot 0.
+                let AbstractDep::External { pod, statement } = &self.shape.dep_edges[s][0] else {
+                    unreachable!("synthetic statement must have a single External dep");
+                };
+                let slot = ext_slot[pod];
+                let stmt = self.external_index.statements[*statement].statement.clone();
+                builder.open_input_st(true, slot, &stmt)?;
+                continue;
+            }
+            let public = !is_output && public_set.contains(&s);
+            if let OperationType::Native(NativeOperation::OpenInputStatement) =
+                &self.operations[s].0
+            {
+                // Staging-time aux carries a `pod_index` from the staging
+                // builder's input slots; re-issue against this POD's
+                // ext_slot mapping.
+                let OperationAux::OpenInputStatement(InputPodOpenStatement { sts_root, .. }) =
+                    &self.operations[s].2
+                else {
+                    unreachable!("OpenInputStatement op without InputPodOpenStatement aux");
+                };
+                let abs_pod = *self
+                    .external_index
+                    .pod_index_by_hash
+                    .get(sts_root)
+                    .expect("staging OpenInputStatement's source pod is in external_index");
+                let slot = ext_slot[&abs_pod];
+                builder.open_input_st(public, slot, &self.statements[s])?;
+                continue;
+            }
+            let wildcards = self
+                .operations_wildcard_values
+                .get(&s)
+                .cloned()
+                .unwrap_or_default();
+            builder.op(public, wildcards, self.operations[s].clone())?;
+        }
+
+        if is_output {
+            for &idx in &self.output_public_indices {
+                let stmt = &self.statements[idx];
+                if builder.statements.iter().any(|(_, s)| s == stmt) {
+                    builder.reveal(stmt)?;
+                } else {
+                    // Output-public produced upstream — open from chain.
+                    builder.open_input_st(true, 0, stmt)?;
+                }
+            }
+        }
+
+        builder.prove(prover).map_err(Error::from)
+    }
+
+    /// For POD `p`: the concrete `MainPod`s to install as inputs (chain
+    /// predecessor at slot 0 if any, externals at slots 1+) plus a map
+    /// from abstract external-pod index to builder slot. The slot map
+    /// drives `open_input_st` calls for synthetics and staging-Opens.
+    fn pod_inputs(
+        &self,
+        p: usize,
+        earlier_pods: &[MainPod],
+    ) -> (Vec<MainPod>, HashMap<usize, usize>) {
+        let mut inputs: Vec<MainPod> = Vec::new();
+        if p > 0 {
+            inputs.push(earlier_pods[p - 1].clone());
+        }
+        let mut refs: BTreeSet<usize> = BTreeSet::new();
+        for &s in &self.output.pod_statements[p] {
+            for dep in &self.shape.dep_edges[s] {
+                if let AbstractDep::External { pod, .. } = dep {
+                    refs.insert(*pod);
+                }
+            }
+        }
+        let mut ext_slot: HashMap<usize, usize> = HashMap::new();
+        for abs_pod in refs {
+            let pod_idx = self.input_pod_idx_by_abs[abs_pod];
+            ext_slot.insert(abs_pod, inputs.len());
+            inputs.push(self.input_pods[pod_idx].clone());
+        }
+        (inputs, ext_slot)
+    }
+}
+
+pub struct MultiPodResult {
+    pub pods: Vec<MainPod>,
+}
+
+impl MultiPodResult {
+    pub fn output_pod(&self) -> &MainPod {
+        self.pods.last().expect("at least one POD")
+    }
+
+    pub fn intermediate_pods(&self) -> &[MainPod] {
+        &self.pods[..self.pods.len() - 1]
+    }
+}
+
+/// Per-POD public-set decisions: which statements each intermediate POD
+/// must reveal so downstream PODs (and the output POD) can reach them
+/// through the chain tree.
+///
+/// Three rules combine:
+/// 1. **Downstream consumption**: any statement consumed by a later POD via
+///    `AbstractDep::Internal` must be public at its producer. This also
+///    covers synthetics whose consumers live downstream (their republish
+///    is exactly this case).
+/// 2. **Output-public flow-up**: any statement in `output_public_indices`
+///    produced upstream of the output POD must be revealed by its
+///    producer so the output POD can `open_input_st` it through slot 0.
+/// 3. **Output POD**: the last POD's set is left empty here. The build
+///    layer handles its fresh public tree separately (no `extend`).
+///
+/// Indexed by POD position. Entry `pod_count - 1` is always empty.
+fn intermediate_public_sets(shape: &InputShape, output: &OutputShape) -> Vec<BTreeSet<usize>> {
+    let pod_count = output.pod_count;
+    let mut sets: Vec<BTreeSet<usize>> = vec![BTreeSet::new(); pod_count];
+    if pod_count == 0 {
+        return sets;
+    }
+
+    let n = shape.num_statements();
+    let mut pod_of: Vec<Option<usize>> = vec![None; n];
+    for (p, stmts) in output.pod_statements.iter().enumerate() {
+        for &s in stmts {
+            pod_of[s] = Some(p);
+        }
+    }
+    let output_pod = pod_count - 1;
+
+    // Rule 1: downstream-consumed statements get revealed by their producer.
+    for (consumer, deps) in shape.dep_edges.iter().enumerate() {
+        let consumer_pod = match pod_of[consumer] {
+            Some(p) => p,
+            None => continue,
+        };
+        for dep in deps {
+            if let AbstractDep::Internal(producer) = dep {
+                if let Some(producer_pod) = pod_of[*producer] {
+                    if producer_pod < consumer_pod {
+                        sets[producer_pod].insert(*producer);
+                    }
+                }
+            }
+        }
+    }
+
+    // Rule 2: output-public statements produced upstream of the output POD.
+    for &s in &shape.output_public_indices {
+        if let Some(producer_pod) = pod_of[s] {
+            if producer_pod < output_pod {
+                sets[producer_pod].insert(s);
+            }
+        }
+    }
+
+    sets
+}
+
+/// Index every public statement of every input POD by content. The dep
+/// graph uses this to recognise when a statement argument refers to an
+/// externally-provided POD's public statement rather than a locally-built
+/// one.
 fn build_external_statement_map(input_pods: &[MainPod]) -> HashMap<Statement, Hash> {
     let mut map = HashMap::new();
     for pod in input_pods {
@@ -650,179 +604,276 @@ fn build_external_statement_map(input_pods: &[MainPod]) -> HashMap<Statement, Ha
     map
 }
 
+/// Convert a concrete `DependencyGraph` into the positional `InputShape`
+/// the solver consumes.
+///
+/// Also runs the external-republish pre-pass: input statements with two
+/// or more downstream consumers are rewritten as synthetic "republish"
+/// statements (zero cost, one `External` dep) appended after the original
+/// statement list. Consumers' deps that pointed at those input statements
+/// become `Internal` references to the synthetic position. Whichever POD
+/// the DP places a synthetic into pays one external-input slot for that
+/// input statement; every downstream consumer reaches it via slot 0
+/// (chain) for one import slot. With multiple consumers, this saves at
+/// least one external-input slot net.
+///
+/// Returns the [`InputShape`] and the side-table [`ExternalIndex`]. Each
+/// synthetic statement's single [`AbstractDep::External`] dep records the
+/// input statement it republishes, so callers can recover the
+/// synthetic-to-statement mapping from `shape.dep_edges` directly.
+fn build_shape_and_index(
+    operations: &[Operation],
+    deps: &DependencyGraph,
+    output_public_indices: &[usize],
+    params: &Params,
+) -> (InputShape, ExternalIndex) {
+    let n_orig = operations.len();
+
+    let mut external_pods: Vec<Hash> = Vec::new();
+    let mut pod_idx: HashMap<Hash, usize> = HashMap::new();
+    let mut external_statements: Vec<ExternalDependency> = Vec::new();
+    let mut external_statement_idx: HashMap<ExternalDependency, usize> = HashMap::new();
+    let mut statement_pod: Vec<usize> = Vec::new();
+
+    for edges in &deps.statement_deps {
+        for src in edges {
+            if let StatementSource::External(ext) = src {
+                if let Entry::Vacant(e) = pod_idx.entry(ext.pod_hash) {
+                    e.insert(external_pods.len());
+                    external_pods.push(ext.pod_hash);
+                }
+                if let Entry::Vacant(e) = external_statement_idx.entry(ext.clone()) {
+                    e.insert(external_statements.len());
+                    external_statements.push(ext.clone());
+                    statement_pod.push(pod_idx[&ext.pod_hash]);
+                }
+            }
+        }
+    }
+
+    // Count downstream consumers per input statement. Dedupe within a
+    // single consumer's dep list (multiple deps on the same input
+    // statement from one consumer still count as one consumer).
+    let mut statement_consumer_count = vec![0_usize; external_statements.len()];
+    for edges in &deps.statement_deps {
+        let mut seen: HashSet<usize> = HashSet::new();
+        for src in edges {
+            if let StatementSource::External(ext) = src {
+                let u = external_statement_idx[ext];
+                if seen.insert(u) {
+                    statement_consumer_count[u] += 1;
+                }
+            }
+        }
+    }
+
+    // Feasibility-driven republish: a statement's POD imports one slot
+    // for chain (if any dep flows through it) plus one slot per distinct
+    // external pod the statement directly references. When the sum
+    // exceeds `max_input_pods`, the POD cannot fit -- we must republish
+    // enough of the statement's input statements so they reach the
+    // consumer through the chain instead. Mark those input statements
+    // here; the synthetic allocation loop below unions this with the
+    // 2+ consumers rule.
+    //
+    // Republishing any input statement forces chain use at the consumer
+    // site, so the post-republish budget is
+    // `(K - R) + 1 <= max_input_pods`, i.e.
+    // `R >= K - max_input_pods + 1` whenever the original
+    // `K + chain_used` would have busted the cap.
+    let max_input_pods = params.max_input_pods;
+    let mut must_republish: Vec<bool> = vec![false; external_statements.len()];
+    for edges in &deps.statement_deps {
+        let mut distinct_pods: BTreeSet<usize> = BTreeSet::new();
+        let mut has_internal = false;
+        let mut statements_by_pod: HashMap<usize, Vec<usize>> = HashMap::new();
+        for src in edges {
+            match src {
+                StatementSource::Internal(_) => has_internal = true,
+                StatementSource::External(ext) => {
+                    let pod = pod_idx[&ext.pod_hash];
+                    let statement = external_statement_idx[ext];
+                    distinct_pods.insert(pod);
+                    statements_by_pod.entry(pod).or_default().push(statement);
+                }
+            }
+        }
+        let k = distinct_pods.len();
+        let chain_already_used = if has_internal { 1 } else { 0 };
+        if k + chain_already_used <= max_input_pods {
+            continue;
+        }
+        // Republish R pods. After republishing, chain is in use, so we
+        // need (K - R) + 1 <= max_input_pods.
+        let r = k + 1 - max_input_pods;
+        // Pick the highest-numbered pods first (deterministic). The
+        // chosen pods drop out of the consumer's direct refs.
+        for pod in distinct_pods.iter().rev().take(r) {
+            for &statement in &statements_by_pod[pod] {
+                must_republish[statement] = true;
+            }
+        }
+    }
+
+    // Opportunistic external-pod bundling: if any input statement from
+    // external pod E is already being republished (multi-consumer or
+    // feasibility), republish all of E's used input statements. The
+    // synth-hosting POD already pays one input-pod slot for E, so
+    // bundling its other input statements is marginal cost there and
+    // frees downstream consumers from re-referencing E. See
+    // `docs/multipod_merkle_statement_tree.md`.
+    let mut bundled_pods: HashSet<usize> = HashSet::new();
+    for u in 0..external_statements.len() {
+        if statement_consumer_count[u] >= 2 || must_republish[u] {
+            bundled_pods.insert(statement_pod[u]);
+        }
+    }
+    for u in 0..external_statements.len() {
+        if bundled_pods.contains(&statement_pod[u]) {
+            must_republish[u] = true;
+        }
+    }
+
+    // Allocate a synthetic statement for each input statement with 2+
+    // consumers OR flagged by the feasibility pre-pass above.
+    // `synthetic_to_statement[i]` is the input-statement index of the
+    // (n_orig + i)-th statement.
+    let mut synthetic_to_statement: Vec<usize> = Vec::new();
+    let mut statement_to_synthetic: Vec<Option<usize>> = vec![None; external_statements.len()];
+    for u in 0..external_statements.len() {
+        if statement_consumer_count[u] >= 2 || must_republish[u] {
+            let synth_idx = n_orig + synthetic_to_statement.len();
+            statement_to_synthetic[u] = Some(synth_idx);
+            synthetic_to_statement.push(u);
+        }
+    }
+    let n_synth = synthetic_to_statement.len();
+
+    // Augmented costs: originals + zero costs for synthetics.
+    let mut costs: Vec<StatementCost> = operations
+        .iter()
+        .map(|op| StatementCost::from_operation(op, params))
+        .collect();
+    costs.extend((0..n_synth).map(|_| StatementCost::default()));
+
+    // Augmented dep_edges. Original statements: External(pod, statement)
+    // becomes Internal(synth_idx) when the input statement is being
+    // republished. Synthetic statements: each has one External dep to
+    // its input statement.
+    let mut dep_edges: Vec<Vec<AbstractDep>> = deps
+        .statement_deps
+        .iter()
+        .map(|edges| {
+            edges
+                .iter()
+                .map(|src| match src {
+                    StatementSource::Internal(d) => AbstractDep::Internal(*d),
+                    StatementSource::External(ext) => {
+                        let u = external_statement_idx[ext];
+                        if let Some(synth_idx) = statement_to_synthetic[u] {
+                            AbstractDep::Internal(synth_idx)
+                        } else {
+                            AbstractDep::External {
+                                pod: pod_idx[&ext.pod_hash],
+                                statement: u,
+                            }
+                        }
+                    }
+                })
+                .collect()
+        })
+        .collect();
+    for &u in &synthetic_to_statement {
+        let ext = &external_statements[u];
+        dep_edges.push(vec![AbstractDep::External {
+            pod: pod_idx[&ext.pod_hash],
+            statement: u,
+        }]);
+    }
+
+    let shape = InputShape {
+        costs,
+        dep_edges,
+        output_public_indices: output_public_indices.to_vec(),
+        num_external_pods: external_pods.len(),
+        statement_pod,
+        params: params.clone(),
+    };
+    let index = ExternalIndex::new(external_pods, external_statements);
+    (shape, index)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::{
-        backends::plonky2::{
-            mock::mainpod::MockProver, primitives::ec::schnorr::SecretKey, signer::Signer,
-        },
-        dict,
-        examples::MOCK_VD_SET,
-        frontend::{Operation as FrontendOp, SignedDictBuilder},
-        lang::load_module,
+        backends::plonky2::mock::mainpod::MockProver, examples::MOCK_VD_SET,
+        frontend::Operation as FrontendOp,
     };
 
     #[test]
-    fn test_single_pod_case() -> Result<()> {
+    fn end_to_end_solve_single_pod() {
         let params = Params::default();
         let vd_set = &*MOCK_VD_SET;
 
         let mut builder = MultiPodBuilder::new(&params, vd_set);
+        builder.pub_op(FrontendOp::eq(1, 1)).expect("op should add");
 
-        // Create a simple signed dict
-        let mut signed_builder = SignedDictBuilder::new(&params);
-        signed_builder.insert("value", 42);
-        let signer = Signer(SecretKey(1u32.into()));
-        let signed_dict = signed_builder.sign(&signer).unwrap();
-
-        // Add operation
-        builder.pub_op(FrontendOp::dict_signed_by(&signed_dict))?;
-
-        // Solve
-        let solved = builder.solve()?;
+        let solved = builder.solve().expect("should solve");
         assert_eq!(solved.solution().pod_count, 1);
+        assert_eq!(solved.solution().pod_statements[0].len(), 1);
+    }
 
-        // Prove
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
+    #[test]
+    fn end_to_end_solve_forces_split() {
+        // Tight max_statements forces splitting into 2 PODs.
+        let params = Params {
+            max_statements: 2,
+            ..Params::default()
+        };
+        let vd_set = &*MOCK_VD_SET;
 
+        let mut builder = MultiPodBuilder::new(&params, vd_set);
+        // 3 independent statements need 2 PODs at max_statements = 2.
+        for i in 0..3_i64 {
+            builder.priv_op(FrontendOp::eq(i, i)).expect("priv op");
+        }
+        builder.pub_op(FrontendOp::eq(100, 100)).expect("pub op");
+
+        let solved = builder.solve().expect("should solve");
+        assert!(solved.solution().pod_count >= 2);
+    }
+
+    #[test]
+    fn prove_single_pod() {
+        let params = Params::default();
+        let vd_set = &*MOCK_VD_SET;
+
+        let mut builder = MultiPodBuilder::new(&params, vd_set);
+        builder.pub_op(FrontendOp::eq(1, 1)).expect("op should add");
+
+        let solved = builder.solve().expect("should solve");
+        assert_eq!(solved.solution().pod_count, 1);
+        let result = solved.prove(&MockProver {}).expect("prove should succeed");
         assert_eq!(result.pods.len(), 1);
-        assert!(result.intermediate_pods().is_empty());
-
-        // Verify the POD
-        result.pods[0].pod.verify().unwrap();
-
-        Ok(())
+        result.pods[0].pod.verify().expect("pod verifies");
     }
 
     #[test]
-    fn test_multi_pod_overflow() -> Result<()> {
-        // Verifies automatic splitting when statements exceed per-POD capacity.
-        //
-        // This test uses independent statements with no dependencies - the only
-        // reason for multiple PODs is the statement limit being exceeded.
+    fn prove_cross_pod_custom_predicate_chain() {
+        // pred_b(X) wraps pred_a(X) wraps Contains(X, "k", 1). With
+        // max_statements = 2, the three resulting statements must split
+        // across PODs, so the output POD's b_out has to reach back
+        // through the chain tree to consume the intermediate pred_a /
+        // Contains it depends on.
+        use crate::{dict, lang::load_module};
+
         let params = Params {
-            max_statements: 6,
-            max_public_statements: 2,
-            // Derived: max_priv_statements = 6 - 2 = 4
-            // With 6 private + 2 public = 8 statements, need ceil(8/4) = 2 PODs
-            max_input_pods: 2,
-            max_input_pods_public_statements: 4,
+            max_statements: 2,
             ..Params::default()
         };
         let vd_set = &*MOCK_VD_SET;
 
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
-
-        // Add 6 independent private statements (no dependencies between them)
-        for i in 0..6i64 {
-            builder.priv_op(FrontendOp::eq(i, i))?;
-        }
-
-        // Add 2 public statements for the output POD
-        builder.pub_op(FrontendOp::eq(100, 100))?;
-        builder.pub_op(FrontendOp::eq(101, 101))?;
-
-        // Solve
-        let solved = builder.solve()?;
-        // 8 statements / 4 per POD = 2 PODs minimum
-        assert!(
-            solved.solution().pod_count >= 2,
-            "Expected at least 2 PODs for 8 statements with max_priv=4, got {}",
-            solved.solution().pod_count
-        );
-        let pod_count = solved.solution().pod_count;
-
-        // Prove and verify
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
-        assert_eq!(result.pods.len(), pod_count);
-
-        for (i, pod) in result.pods.iter().enumerate() {
-            pod.pod
-                .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_duplicate_statement_content_reused_within_pod() -> Result<()> {
-        // This test verifies that duplicate statement content is reused within a POD.
-        // We run three operations, but each produces the same statement. This allows us to
-        // deuplicate the private statement, matching the solver's deduplication logic.
-        // Since there is only space for 2 private statements with these parameters, the
-        // test can only succeed if deduplication is working correctly.
-        // Public statements/reveals of private statements are not deduplicated, so we can
-        // have 3 of them.
-        let params = Params {
-            max_statements: 5,
-            max_public_statements: 3,
-            // Derived: max_priv_statements = 2
-            max_input_pods: 2,
-            max_input_pods_public_statements: 4,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
-        builder.pub_op(FrontendOp::eq(7, 7))?;
-        builder.pub_op(FrontendOp::eq(7, 7))?;
-        builder.pub_op(FrontendOp::eq(7, 7))?;
-
-        let solved = builder.solve()?;
-        let pod_count = solved.solution().pod_count;
-
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
-        assert_eq!(result.pods.len(), pod_count);
-        for (i, pod) in result.pods.iter().enumerate() {
-            pod.pod
-                .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_cross_pod_dependencies() -> Result<()> {
-        // Verifies that a dependency chain can be split across PODs.
-        //
-        // This tests the core multi-POD capability: when a dependency chain is too
-        // long to fit in the output POD, intermediate statements must be proved in
-        // earlier PODs and made public so the output POD can access them.
-        //
-        // Chain: b_out -> a_out -> contains
-        //   - contains: base statement (dict_contains)
-        //   - a_out: custom predicate taking contains as argument
-        //   - b_out: custom predicate taking a_out as argument (OUTPUT-PUBLIC)
-        //
-        // With max_priv_statements = 2, we can't fit all 3 in one POD.
-        // Expected solution:
-        //   - POD 0 (intermediate): contains, a_out (with a_out public)
-        //   - POD 1 (output): copy(a_out), b_out
-        //
-        // This requires intermediate PODs to feed INTO the output POD.
-
-        // Tight params to force the dependency chain to be split.
-        // With max_priv_statements = 2, we can't fit contains + a_out + b_out's
-        // dependencies all in one POD.
-        let params = Params {
-            max_statements: 4,
-            max_public_statements: 2,
-            // max_priv_statements = 2
-            max_input_pods: 4,
-            max_input_pods_public_statements: 20,
-            max_custom_predicate_verifications: 10,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        // pred_a accepts a Contains statement
-        // pred_b accepts a pred_a statement (Custom statement from pred_a)
         let module = load_module(
             r#"
             pred_a(X) = AND(Contains(X, "k", 1))
@@ -837,908 +888,1139 @@ mod tests {
 
         let mut builder = MultiPodBuilder::new(&params, vd_set);
 
-        // Statement 0: Contains (base of the chain)
         let dict = dict!({"k" => 1});
-        let contains = builder.priv_op(FrontendOp::dict_contains(dict, "k", 1))?;
+        let contains = builder
+            .priv_op(FrontendOp::dict_contains(dict, "k", 1))
+            .expect("contains");
+        let a_out = builder
+            .priv_op(FrontendOp::custom(
+                batch.predicate_ref_by_name("pred_a").unwrap(),
+                [contains],
+            ))
+            .expect("pred_a");
+        let _b_out = builder
+            .pub_op(FrontendOp::custom(
+                batch.predicate_ref_by_name("pred_b").unwrap(),
+                [a_out],
+            ))
+            .expect("pred_b");
 
-        // Statement 1: Custom(pred_a), depends on contains
-        let a_out = builder.priv_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_a").unwrap(),
-            [contains],
-        ))?;
-
-        // Statement 2: Custom(pred_b), depends on a_out - make this output-public
-        // This forces the dependency chain to be resolved for the output POD.
-        let _b_out = builder.pub_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_b").unwrap(),
-            [a_out],
-        ))?;
-
-        // Solve - this finds a multi-POD solution where intermediate PODs
-        // provide dependencies to the output POD.
-        let solved = builder.solve()?;
-        let solution = solved.solution();
-
-        // Expected: exactly 2 PODs
-        //   Solution A:
-        //   - POD 0 (intermediate): public statements 0 (contains)
-        //   - POD 1 (output): inherits statement 0 (contains) from POD0, statement 1 (a_out),
-        //   public statement 2 (b_out)
-        //   Solution B:
-        //   - POD 0 (intermediate): statements 0 (contains), public statement 1 (a_out)
-        //   - POD 1 (output): inherits statement 1 (a_out) from POD0, public statement 2 (b_out)
-
-        // Statement 2 (b_out) should be public in POD 1 (it's output-public)
-        assert!(
-            solution.pod_public_statements[1].contains(&2),
-            "Statement 2 (b_out) should be public in output POD"
-        );
-
-        // Prove and verify all PODs
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
-
+        let solved = builder.solve().expect("should solve");
+        assert!(solved.solution().pod_count >= 2);
+        let result = solved.prove(&MockProver {}).expect("prove should succeed");
         for (i, pod) in result.pods.iter().enumerate() {
             pod.pod
                 .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
+                .unwrap_or_else(|e| panic!("POD {} verification failed: {:?}", i, e));
         }
-
-        Ok(())
     }
 
     #[test]
-    fn test_isolated_pods_when_no_inputs_allowed() -> Result<()> {
-        // Verifies that PODs are completely isolated when max_input_pods = 0.
-        //
-        // With no input PODs allowed, each generated POD must independently prove
-        // all statements it contains - it cannot reference earlier PODs.
-        // This is an edge case but validates the input POD constraint.
+    fn prove_chain_two_pods() {
+        // Tight max_statements forces splitting; verifies that chain
+        // monotonicity (extend_input_pod0_public_statements) carries the
+        // intermediate POD's statement forward so the output POD can
+        // reach it.
         let params = Params {
-            max_statements: 4,
-            max_public_statements: 2,
-            // Derived: max_priv_statements = 4 - 2 = 2
-            max_input_pods: 0, // No input pods allowed - each POD is isolated
-            max_input_pods_public_statements: 0,
+            max_statements: 2,
             ..Params::default()
         };
         let vd_set = &*MOCK_VD_SET;
 
         let mut builder = MultiPodBuilder::new(&params, vd_set);
-
-        // Add 4 independent private statements (no dependencies)
-        // With max_priv=2, need 2 PODs. Since max_input_pods=0, they can't share.
-        for i in 0..4i64 {
-            builder.priv_op(FrontendOp::eq(i, i))?;
+        for i in 0..3_i64 {
+            builder.priv_op(FrontendOp::eq(i, i)).expect("priv op");
         }
+        builder.pub_op(FrontendOp::eq(100, 100)).expect("pub op");
 
-        // Add 2 public statements for the output POD
-        builder.pub_op(FrontendOp::eq(100, 100))?;
-        builder.pub_op(FrontendOp::eq(101, 101))?;
-
-        let solved = builder.solve()?;
-
-        // 6 statements / 2 per POD = 3 PODs minimum
-        assert!(
-            solved.solution().pod_count >= 2,
-            "Expected at least 2 PODs, got {}",
-            solved.solution().pod_count
-        );
-
-        // Prove and verify
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
-
+        let solved = builder.solve().expect("should solve");
+        assert!(solved.solution().pod_count >= 2);
+        let result = solved.prove(&MockProver {}).expect("prove should succeed");
         for (i, pod) in result.pods.iter().enumerate() {
             pod.pod
                 .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
+                .unwrap_or_else(|e| panic!("POD {} verification failed: {:?}", i, e));
         }
-
-        Ok(())
     }
 
     #[test]
-    fn test_zero_public_capacity_fails() {
-        // Test that setting max_public_statements = 0 with a public operation
-        // results in a solver error (infeasible configuration).
-        let params = Params {
-            max_statements: 10,
-            max_public_statements: 0, // No public statements allowed
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
+    fn prove_with_external_pod_arg() {
+        // User-adds an external pod and passes one of its public
+        // statements directly as an op arg (via a custom predicate that
+        // wraps it). Exercises the staging auto-import path:
+        // `MainPodBuilder::ensure_statement` emits an
+        // `OpenInputStatement` op at staging time, the dep graph models
+        // it as an External-dep node, and replay re-issues the open
+        // against the per-POD builder's slot.
+        use crate::lang::load_module;
 
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
-
-        // Try to add a public operation
-        let _ = builder.pub_op(FrontendOp::eq(1, 1));
-
-        // Solving should fail because we can't satisfy the public statement requirement
-        let result = builder.solve();
-        assert!(
-            result.is_err(),
-            "Expected solver to fail with zero public capacity, but it succeeded"
-        );
-    }
-
-    #[test]
-    fn test_max_pods_exceeded_error() {
-        // Test that exceeding max_pods gives a clear error message.
-        // With max_statements=3 and max_public_statements=1, we have
-        // max_priv_statements = 2. So 10 statements requires 5 PODs.
-        let params = Params {
-            max_statements: 3,
-            max_public_statements: 1,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        // Set max_pods to 2, which is less than the 5 PODs needed
-        let options = Options { max_pods: 2 };
-        let mut builder = MultiPodBuilder::new_with_options(&params, vd_set, options);
-
-        // Add 10 statements (requires 5 PODs). First one is public (required).
-        let _ = builder.pub_op(FrontendOp::eq(0, 0));
-        for i in 1..10 {
-            let _ = builder.priv_op(FrontendOp::eq(i, i));
-        }
-
-        // Solving should fail with a clear error about max_pods
-        let result = builder.solve();
-        assert!(
-            result.is_err(),
-            "Expected solver to fail when max_pods exceeded"
-        );
-
-        let err_msg = result.unwrap_err().to_string();
-        assert!(
-            err_msg.contains("requires at least") && err_msg.contains("PODs"),
-            "Error message should explain POD requirement: {}",
-            err_msg
-        );
-        assert!(
-            err_msg.contains("Options::max_pods"),
-            "Error message should suggest increasing Options::max_pods: {}",
-            err_msg
-        );
-    }
-
-    #[test]
-    fn test_external_pods_only_added_where_needed() -> Result<()> {
-        // Verifies that external input PODs are only added to generated PODs
-        // that actually need them based on statement dependencies.
-        //
-        // Setup:
-        // - Two external PODs: ext_A and ext_B, each with a public statement
-        // - max_input_pods = 1 (each generated POD can only have 1 input POD)
-        // - Private statements that copy from different external PODs force overflow
-        //
-        // With max_input_pods = 1, this only works if each generated POD
-        // includes only the external POD it actually depends on.
-
-        let params = Params {
-            max_statements: 4,        // Small limit
-            max_public_statements: 2, // max_priv_statements = 4 - 2 = 2
-            max_input_pods: 1,        // Only 1 input POD allowed per generated POD
-            max_input_pods_public_statements: 4,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        // Create external POD A with a public statement
-        let prover = MockProver {};
-        let mut builder_a = MainPodBuilder::new(&params, vd_set);
-        builder_a.pub_op(FrontendOp::eq(100, 100))?;
-        let ext_pod_a = builder_a.prove(&prover)?;
-
-        // Create external POD B with a public statement
-        let mut builder_b = MainPodBuilder::new(&params, vd_set);
-        builder_b.pub_op(FrontendOp::eq(200, 200))?;
-        let ext_pod_b = builder_b.prove(&prover)?;
-
-        // Get the actual statements from the proved PODs
-        let stmt_a = ext_pod_a
-            .pod
-            .pub_statements()
-            .into_iter()
-            .find(|s| !s.is_none())
-            .expect("ext_pod_a should have a public statement");
-        let stmt_b = ext_pod_b
-            .pod
-            .pub_statements()
-            .into_iter()
-            .find(|s| !s.is_none())
-            .expect("ext_pod_b should have a public statement");
-
-        // Create MultiPodBuilder and add both external PODs
-        let mut multi_builder = MultiPodBuilder::new(&params, vd_set);
-        multi_builder.add_pod(ext_pod_a.clone())?;
-        multi_builder.add_pod(ext_pod_b.clone())?;
-
-        // Add private operations that reference different external PODs.
-        // These will force multiple PODs due to private statement limits.
-        multi_builder.priv_op(FrontendOp::copy(stmt_a))?;
-        multi_builder.priv_op(FrontendOp::eq(101, 101))?;
-        multi_builder.priv_op(FrontendOp::copy(stmt_b))?;
-        multi_builder.priv_op(FrontendOp::eq(201, 201))?;
-
-        // Add 2 public statements (within single output POD limit)
-        multi_builder.pub_op(FrontendOp::eq(300, 300))?;
-        multi_builder.pub_op(FrontendOp::eq(301, 301))?;
-
-        // With 6 statements and max_priv_statements = 2, we need multiple PODs.
-        // Each POD should only include the external POD it depends on.
-
-        let solved = multi_builder.solve()?;
-        assert!(
-            solved.solution().pod_count >= 2,
-            "Expected at least 2 PODs, got {}",
-            solved.solution().pod_count
-        );
-
-        let result = solved.prove(&prover)?;
-
-        // Verify all PODs
-        for (i, pod) in result.pods.iter().enumerate() {
-            pod.pod
-                .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_private_statement_not_leaked_to_output_pod() -> Result<()> {
-        // Verifies that private statements do not appear in the output POD's public slots.
-        // The solver enforces that only user-requested public statements can be
-        // public in the output POD (the last POD).
-
-        let params = Params {
-            max_statements: 4,
-            max_public_statements: 2,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
-
-        // Add private statements (indices 0, 1, 2) - should NOT appear in output POD public slots
-        builder.priv_op(FrontendOp::eq(100, 100))?;
-        builder.priv_op(FrontendOp::eq(101, 101))?;
-        builder.priv_op(FrontendOp::eq(102, 102))?;
-
-        // Add public statements (indices 3, 4) - these SHOULD appear in output POD public slots
-        builder.pub_op(FrontendOp::eq(200, 200))?;
-        builder.pub_op(FrontendOp::eq(201, 201))?;
-
-        let solved = builder.solve()?;
-        let solution = solved.solution();
-
-        // Check that the output POD's public statements are exactly the user-requested public ones.
-        // The output POD is always the last one (index pod_count - 1).
-        let output_pod_idx = solution.pod_count - 1;
-        let output_public = &solution.pod_public_statements[output_pod_idx];
-        assert!(
-            output_public.contains(&3),
-            "Public statement 3 should be public in output POD"
-        );
-        assert!(
-            output_public.contains(&4),
-            "Public statement 4 should be public in output POD"
-        );
-
-        // Private statements should NOT be public in output POD
-        assert!(
-            !output_public.contains(&0),
-            "Private statement 0 should NOT be public in output POD"
-        );
-        assert!(
-            !output_public.contains(&1),
-            "Private statement 1 should NOT be public in output POD"
-        );
-        assert!(
-            !output_public.contains(&2),
-            "Private statement 2 should NOT be public in output POD"
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_too_many_public_statements_error() -> Result<()> {
-        // Verifies that requesting more public statements than max_public_statements
-        // results in a clear error (since all public statements must fit in one output POD).
-
-        let params = Params {
-            max_statements: 10,
-            max_public_statements: 2,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
-
-        // Add 3 public statements, but max is 2
-        builder.pub_op(FrontendOp::eq(1, 1))?;
-        builder.pub_op(FrontendOp::eq(2, 2))?;
-        builder.pub_op(FrontendOp::eq(3, 3))?;
-
-        let result = builder.solve();
-
-        assert!(result.is_err());
-        let err_msg = result.unwrap_err().to_string();
-        assert!(
-            err_msg.contains("Too many public statements"),
-            "Expected 'Too many public statements' error, got: {}",
-            err_msg
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_external_pods_counted_in_input_limit() -> Result<()> {
-        // Verifies that external input PODs are counted toward max_input_pods while
-        // still allowing the solver to route external premises through intermediate PODs.
-        //
-        // Setup:
-        // - max_input_pods = 2
-        // - 3 external PODs (A, B, C), each with a public statement
-        // - 3 public operations, each copying from a different external POD
-        //
-        // A direct 1-POD layout would need 3 external inputs in the output POD (infeasible),
-        // so the solver should split the work and keep each generated POD within input limits.
-
-        let params = Params {
-            max_statements: 10,
-            max_public_statements: 5,
-            max_input_pods: 2, // Only 2 input PODs allowed per generated POD
-            max_input_pods_public_statements: 10,
-            ..Params::default()
-        };
+        let params = Params::default();
         let vd_set = &*MOCK_VD_SET;
         let prover = MockProver {};
 
-        // Create 3 external PODs, each with a distinct public statement
-        let mut builder_a = MainPodBuilder::new(&params, vd_set);
-        builder_a.pub_op(FrontendOp::eq(100, 100))?;
-        let ext_pod_a = builder_a.prove(&prover)?;
+        let module = load_module(
+            r#"
+            pred_a(X) = AND(Lt(X, 2))
+            "#,
+            "test",
+            &params,
+            &[],
+        )
+        .expect("load module");
+        let batch = &module.batch;
 
-        let mut builder_b = MainPodBuilder::new(&params, vd_set);
-        builder_b.pub_op(FrontendOp::eq(200, 200))?;
-        let ext_pod_b = builder_b.prove(&prover)?;
-
-        let mut builder_c = MainPodBuilder::new(&params, vd_set);
-        builder_c.pub_op(FrontendOp::eq(300, 300))?;
-        let ext_pod_c = builder_c.prove(&prover)?;
-
-        // Get the actual statements from the proved PODs
-        let stmt_a = ext_pod_a
-            .pod
-            .pub_statements()
-            .into_iter()
-            .find(|s| !s.is_none())
-            .expect("ext_pod_a should have a public statement");
-        let stmt_b = ext_pod_b
-            .pod
-            .pub_statements()
-            .into_iter()
-            .find(|s| !s.is_none())
-            .expect("ext_pod_b should have a public statement");
-        let stmt_c = ext_pod_c
-            .pod
-            .pub_statements()
-            .into_iter()
-            .find(|s| !s.is_none())
-            .expect("ext_pod_c should have a public statement");
-
-        // Create MultiPodBuilder and add all 3 external PODs
-        let mut multi_builder = MultiPodBuilder::new(&params, vd_set);
-        multi_builder.add_pod(ext_pod_a)?;
-        multi_builder.add_pod(ext_pod_b)?;
-        multi_builder.add_pod(ext_pod_c)?;
-
-        // Add public operations that each depend on a different external POD.
-        multi_builder.pub_op(FrontendOp::copy(stmt_a))?;
-        multi_builder.pub_op(FrontendOp::copy(stmt_b))?;
-        multi_builder.pub_op(FrontendOp::copy(stmt_c))?;
-
-        // Solver should find a feasible multi-POD layout that respects input limits.
-        let solved = multi_builder.solve()?;
-        assert!(
-            solved.solution().pod_count >= 2,
-            "Expected at least 2 PODs to satisfy max_input_pods=2 with 3 external sources"
-        );
-
-        let result = solved.prove(&prover)?;
-        for (i, pod) in result.pods.iter().enumerate() {
-            pod.pod
-                .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_explicit_contains_not_double_counted_as_anchored_key() -> Result<()> {
-        // Verifies that when a Contains statement is explicitly added and then used
-        // as an anchored key argument, it's not double-counted in statement limits.
-        //
-        // Background: MainPodBuilder auto-inserts Contains statements for anchored keys
-        // (dict, key pairs used as arguments to gt(), eq(), etc.). But if the Contains
-        // was already explicitly added, no auto-insertion happens (PR 456).
-        //
-        // The solver must NOT count anchored key overhead when the producing Contains
-        // statement is already in the same POD.
-        //
-        // Setup:
-        // - max_priv_statements = 4
-        // - Statement 0: dict_contains (public) - produces anchored key (dict, "x")
-        // - Statements 1, 2, 3: gt(stmt_0, val) - each references the anchored key
-        //
-        // Correct counting for single POD:
-        // - stmt_sum = 4 (statements 0-3)
-        // - anchored_key_sum = 0 (statement 0 already provides the anchored key)
-        // - Total = 4 ≤ max_priv_statements ✓
-        //
-        // Incorrect (double-counting) would give:
-        // - stmt_sum = 4 + anchored_key_sum = 1 → Total = 5 > 4 ✗
-
-        let params = Params {
-            max_statements: 5,
-            max_public_statements: 1, // max_priv_statements = 5 - 1 = 4
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
-
-        // Statement 0: public Contains - produces anchored key (dict, "x")
-        let dict = dict!({"x" => 100});
-        let contains_stmt = builder.pub_op(FrontendOp::dict_contains(dict, "x", 100))?;
-
-        // Statements 1, 2, 3: each uses contains_stmt as an anchored key
-        builder.priv_op(FrontendOp::gt(contains_stmt.clone(), 0))?;
-        builder.priv_op(FrontendOp::gt(contains_stmt.clone(), 1))?;
-        builder.priv_op(FrontendOp::gt(contains_stmt, 2))?;
-
-        // With correct counting, all 4 statements fit in 1 POD
-        let solved = builder.solve()?;
-        assert_eq!(
-            solved.solution().pod_count,
-            1,
-            "All statements should fit in 1 POD when Contains is not double-counted. \
-             Got {} PODs, which suggests the explicit Contains is being incorrectly \
-             counted as both a statement AND an anchored key overhead.",
-            solved.solution().pod_count
-        );
-
-        // Verify proving works
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
-        assert_eq!(result.pods.len(), 1);
-
-        result.output_pod().pod.verify().unwrap();
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_anchored_key_overhead_counted_in_statement_limit() -> Result<()> {
-        // Verifies that anchored key overhead is correctly counted toward statement limits.
-        //
-        // When a Contains statement is used as an argument to operations like gt(),
-        // it creates an "anchored key" reference. If the gt() is proved in a different
-        // POD than the original Contains, MainPodBuilder auto-inserts a local Contains
-        // statement for that anchored key. The solver must account for this overhead.
-        //
-        // Setup:
-        // - max_priv_statements = 4 (small limit)
-        // - Statement A: dict_contains (public, in POD 0)
-        // - Statement B: eq (public, in POD 0)
-        // - Statements C, D, E: gt(A, val) - each uses A as an anchored key
-        //
-        // The solver must account for the anchored key Contains statements that will
-        // be auto-inserted when gt operations are proved in PODs other than POD 0.
-
-        let params = Params {
-            max_statements: 6,
-            max_public_statements: 2, // max_priv_statements = 6 - 2 = 4
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
-
-        // Statement A: public Contains - proved in POD 0
-        let dict = dict!({"x" => 100});
-        let stmt_a = builder.pub_op(FrontendOp::dict_contains(dict, "x", 100))?;
-
-        // Statement B: another public statement in POD 0
-        builder.pub_op(FrontendOp::eq(200, 200))?;
-
-        // Statements C, D, E: each uses stmt_a as an anchored key
-        // When proved in a different POD, each needs a local Contains for the anchored key
-        builder.priv_op(FrontendOp::gt(stmt_a.clone(), 0))?;
-        builder.priv_op(FrontendOp::gt(stmt_a.clone(), 1))?;
-        builder.priv_op(FrontendOp::gt(stmt_a, 2))?;
-
-        let prover = MockProver {};
-        let result = builder.solve()?.prove(&prover)?;
-
-        // Verify all PODs
-        for (i, pod) in result.pods.iter().enumerate() {
-            pod.pod
-                .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_mixed_internal_and_external_pods_work_within_limit() -> Result<()> {
-        // Verifies that scenarios with both internal and external dependencies work
-        // when the total input count stays within max_input_pods.
-        //
-        // Setup:
-        // - 1 external POD with a public statement
-        // - 2 public dict_contains statements (uses anchored keys)
-        // - 2 private gt statements that reference the dict_contains via anchored keys
-        // - 1 private copy of the external POD's statement
-        //
-        // This tests that mixing internal POD dependencies (from earlier generated PODs)
-        // and external POD dependencies (from user-provided input PODs) works correctly.
-
-        let params = Params {
-            max_statements: 10,
-            max_public_statements: 3, // max_priv_statements = 7
-            max_input_pods: 3,        // Allow up to 3 inputs per POD
-            max_input_pods_public_statements: 10,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
-        let prover = MockProver {};
-
-        // Create 1 external POD
         let mut ext_builder = MainPodBuilder::new(&params, vd_set);
-        ext_builder.pub_op(FrontendOp::eq(9999, 9999))?;
-        let ext_pod = ext_builder.prove(&prover)?;
-
-        let stmt_ext = ext_pod
+        ext_builder
+            .pub_op(FrontendOp::lt(1, 2))
+            .expect("ext pub op");
+        let ext_pod = ext_builder.prove(&prover).expect("ext prove");
+        let ext_stmt = ext_pod
             .pod
             .pub_statements()
             .into_iter()
             .find(|s| !s.is_none())
-            .expect("ext_pod should have a public statement");
+            .expect("ext pod has a public statement");
 
         let mut builder = MultiPodBuilder::new(&params, vd_set);
-        builder.add_pod(ext_pod)?;
+        builder.add_pod(ext_pod).expect("add ext pod");
+        builder
+            .pub_op(FrontendOp::custom(
+                batch.predicate_ref_by_name("pred_a").unwrap(),
+                [ext_stmt],
+            ))
+            .expect("pred_a on ext stmt");
 
-        // Output POD: public Contains statements
-        let dict0 = dict!({"x" => 100});
-        let dict1 = dict!({"y" => 200});
-        let contains_0 = builder.pub_op(FrontendOp::dict_contains(dict0, "x", 100))?;
-        let contains_1 = builder.pub_op(FrontendOp::dict_contains(dict1, "y", 200))?;
-
-        // Statements that depend on output POD
-        builder.priv_op(FrontendOp::gt(contains_0, 0))?;
-        builder.priv_op(FrontendOp::gt(contains_1, 0))?;
-
-        // Depend on external POD
-        builder.priv_op(FrontendOp::copy(stmt_ext))?;
-
-        // This should succeed - total inputs per POD should stay within limit
-        let result = builder.solve()?.prove(&prover)?;
-
+        let solved = builder.solve().expect("should solve");
+        let result = solved.prove(&prover).expect("prove should succeed");
+        assert!(!result.pods.is_empty());
         for (i, pod) in result.pods.iter().enumerate() {
             pod.pod
                 .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
+                .unwrap_or_else(|e| panic!("POD {} verification failed: {:?}", i, e));
         }
-
-        Ok(())
     }
 
+    /// Captured `multi_pod_tree::InputShape` for zk-craft's
+    /// `CraftRefineryCracked` action (episode-1 plugin). Used to
+    /// stress-test the partitioner against a realistic, large input.
+    /// See `sdk/src/tests.rs::capture_cracked_refinery_input_shape`
+    /// in the zk-craft repo for how the fixture is regenerated.
+    const CRACKED_REFINERY_FIXTURE: &str = include_str!("tests/fixtures/cracked_refinery.json");
+
+    /// Pins partition quality on the cracked-refinery fixture. The
+    /// resource-induced lower bound (`max_r ceil(total_r / cap_r)` over
+    /// per-statement-summable resources) is K=12, driven by 96
+    /// custom-predicate verifications at cap 8/POD. Under statement-
+    /// table accounting the heuristic reaches K=14, leaving a 2-POD gap
+    /// to MILP's K=12 optimum. See
+    /// `docs/multipod_merkle_statement_tree.md` ("Custom predicate
+    /// body+head locality" and "Future generators under consideration")
+    /// for the structural reason and candidates that might close it.
     #[test]
-    fn test_signed_by_limit_forces_multi_pod() -> Result<()> {
-        // Verifies that the solver respects max_signed_by per POD (C6f).
-        //
-        // Setup:
-        // - max_signed_by = 2 (small limit)
-        // - 4 SignedBy operations
-        // - Other limits high enough not to interfere
-        //
-        // Expected: Solver creates exactly 2 PODs since 4 SignedBy / 2 per POD = 2 PODs
-        let params = Params {
-            max_statements: 48,
-            max_public_statements: 8,
-            // Derived: max_priv_statements = 48 - 8 = 40 (plenty of room)
-            max_signed_by: 2, // Small limit to force splitting
-            max_input_pods: 10,
-            max_input_pods_public_statements: 20,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
+    fn cracked_refinery_fixture_partitions() {
+        let shape: InputShape =
+            serde_json::from_str(CRACKED_REFINERY_FIXTURE).expect("fixture deserializes");
+        assert_eq!(shape.num_statements(), 345);
 
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
+        let outcome = partition::partition(&shape);
+        match outcome {
+            Some(out) => {
+                eprintln!(
+                    "cracked refinery partitioned into {} PODs, sizes={:?}",
+                    out.pod_count,
+                    out.pod_statements
+                        .iter()
+                        .map(|v| v.len())
+                        .collect::<Vec<_>>()
+                );
+                let breakdown = diagnostics::SolutionBreakdown::from_solution(&shape, &out);
+                eprintln!("{}", breakdown);
+                let total_placed: usize = out.pod_statements.iter().map(|v| v.len()).sum();
+                assert_eq!(total_placed, shape.num_statements());
+                assert_eq!(
+                    out.pod_count, 14,
+                    "expected 14 PODs after bin-packing's can_extend was \
+                     taught to charge import slots against max_statements; \
+                     MILP optimum is 12"
+                );
 
-        // Create 4 different signed dicts
-        for i in 0..4i64 {
-            let mut signed_builder = SignedDictBuilder::new(&params);
-            signed_builder.insert("id", i);
-            let signer = Signer(SecretKey((i as u32 + 1).into()));
-            let signed_dict = signed_builder.sign(&signer).unwrap();
-            builder.priv_op(FrontendOp::dict_signed_by(&signed_dict))?;
+                // Probe whether the DP layer beats greedy on bin-packing's
+                // ordering for this fixture. Pre-statement-table-accounting
+                // the parity sweep recorded `K_bp_dp < K_bp_greedy` = 0;
+                // under the new cap that still holds on cracked-refinery.
+                // Pinned to catch any future divergence.
+                let identity: Vec<usize> = (0..shape.num_statements()).collect();
+                let bp_ordering =
+                    partition::kahn_bin_packing(&shape, &identity).expect("DAG must be acyclic");
+                let k_greedy = partition::simulate_greedy_k(&shape, &bp_ordering)
+                    .expect("greedy partition must be feasible on bin-packing's ordering");
+                let k_dp = partition::partition_with_ordering(&shape, &bp_ordering)
+                    .expect("DP must find a feasible partition on bin-packing's ordering")
+                    .pod_count;
+                eprintln!(
+                    "K_bp_greedy={} K_bp_dp={} on bin-packing's ordering",
+                    k_greedy, k_dp
+                );
+                assert_eq!(
+                    k_greedy, k_dp,
+                    "DP and greedy diverge on bin-packing's cracked-refinery \
+                     ordering; the DP layer earned its place — update doc."
+                );
+            }
+            None => {
+                let diag = diagnostics::diagnose_failure(&shape);
+                panic!(
+                    "partition() returned None; diagnosis: {}",
+                    diag.as_ref()
+                        .map(|v| v.to_string())
+                        .unwrap_or_else(|| "None".to_string())
+                );
+            }
         }
-
-        // Add one public statement for output
-        builder.pub_op(FrontendOp::eq(100, 100))?;
-
-        let solved = builder.solve()?;
-        // 4 SignedBy / 2 per POD = exactly 2 PODs
-        assert_eq!(
-            solved.solution().pod_count,
-            2,
-            "Expected exactly 2 PODs for 4 SignedBy with max_signed_by=2, got {}",
-            solved.solution().pod_count
-        );
-        let pod_count = solved.solution().pod_count;
-
-        // Prove and verify
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
-        assert_eq!(result.pods.len(), pod_count);
-
-        for (i, pod) in result.pods.iter().enumerate() {
-            pod.pod
-                .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
-        }
-
-        Ok(())
     }
 
+    /// Probe: what is MILP's K on cracked-refinery under the current
+    /// caps? Last measured under statement-table accounting: K=11
+    /// infeasible, K=12 feasible — so MILP optimum is K=12, unchanged
+    /// from before the accounting fix. `#[ignore]`d because SCIP takes
+    /// ~23s. Kept as a regression check: if the fixture or caps change,
+    /// re-running this probe re-establishes the optimum so any new
+    /// heuristic gap can be measured against it.
     #[test]
-    fn test_long_dependency_chain_spans_multiple_pods() -> Result<()> {
-        // Verifies that a long dependency chain correctly cascades through multiple
-        // intermediate PODs before reaching the output POD.
-        //
-        // Chain: d_out -> c_out -> b_out -> a_out -> contains (5 statements)
-        //
-        // With max_priv_statements = 2, each POD can hold at most 2 statements.
-        // Cross-POD dependencies are available via input PODs without needing copies.
-        // Expected solution with 3 PODs (ceil(5/2) = 3):
-        //   - POD 0 (intermediate): contains, a_out (a_out public for POD 1)
-        //   - POD 1 (intermediate): b_out, c_out (c_out public for POD 2)
-        //   - POD 2 (output): d_out (public)
+    #[ignore]
+    fn cracked_refinery_fixture_milp() {
+        let shape: InputShape =
+            serde_json::from_str(CRACKED_REFINERY_FIXTURE).expect("fixture deserializes");
+        for k in [11, 12] {
+            let out = partition_milp::solve_for_k(&shape, k);
+            match out {
+                Some(o) => eprintln!(
+                    "MILP K={}: feasible, sizes={:?}",
+                    k,
+                    o.pod_statements.iter().map(|v| v.len()).collect::<Vec<_>>()
+                ),
+                None => eprintln!("MILP K={}: no solution (infeasible or timed out)", k),
+            }
+        }
+    }
 
-        let params = Params {
-            max_statements: 4,
-            max_public_statements: 2,
-            // max_priv_statements = 2
-            max_input_pods: 4,
-            max_input_pods_public_statements: 20,
-            max_custom_predicate_verifications: 10,
-            ..Params::default()
-        };
-        let vd_set = &*MOCK_VD_SET;
+    /// Probe: can the DP find K=12 on cracked-refinery if we feed it the
+    /// "right" ordering derived from MILP's K=12 partition? Builds an
+    /// ordering by sorting statements by (pod_index in MILP's solution,
+    /// position in a baseline topological sort), so each MILP POD is a
+    /// contiguous block. If DP returns K=12, the ordering layer is the
+    /// whole gap (the DP can hit the optimum if generation produces this
+    /// layout). If DP returns K>12, something else is going on.
+    /// `#[ignore]`'d because it calls SCIP.
+    #[test]
+    #[ignore]
+    fn cracked_refinery_milp_ordering_probe() {
+        let shape: InputShape =
+            serde_json::from_str(CRACKED_REFINERY_FIXTURE).expect("fixture deserializes");
+        let milp_out = partition_milp::solve_for_k(&shape, 12)
+            .expect("MILP must find K=12 (regression in fixture or MILP formulation otherwise)");
 
-        // Chain of predicates: each accepts the output of the previous
-        let module = load_module(
-            r#"
-            pred_a(X) = AND(Contains(X, "k", 1))
-            pred_b(X) = AND(pred_a(X))
-            pred_c(X) = AND(pred_b(X))
-            pred_d(X) = AND(pred_c(X))
-            "#,
-            "test",
-            &params,
-            &[],
-        )
-        .expect("load module");
-        let batch = &module.batch;
+        let n = shape.num_statements();
+        let mut pod_of: Vec<usize> = vec![usize::MAX; n];
+        for (p, stmts) in milp_out.pod_statements.iter().enumerate() {
+            for &s in stmts {
+                pod_of[s] = p;
+            }
+        }
 
-        let mut builder = MultiPodBuilder::new(&params, vd_set);
+        // Baseline global topo order. Within each MILP POD, we'll keep
+        // statements in this order; across PODs, the lower POD index
+        // comes first. The MILP's topo precedence constraint guarantees
+        // this concatenation is itself a valid topological ordering.
+        let identity_priority: Vec<usize> = (0..n).collect();
+        let global_topo = partition::kahn_with_priority(&shape, &identity_priority)
+            .expect("statement DAG must be acyclic");
+        let mut pos_in_global = vec![0_usize; n];
+        for (i, &s) in global_topo.iter().enumerate() {
+            pos_in_global[s] = i;
+        }
 
-        // Build the chain: contains -> a_out -> b_out -> c_out -> d_out
-        let dict = dict!({"k" => 1});
-        let contains = builder.priv_op(FrontendOp::dict_contains(dict, "k", 1))?;
+        let mut ordering: Vec<usize> = (0..n).collect();
+        ordering.sort_by_key(|&s| (pod_of[s], pos_in_global[s]));
 
-        let a_out = builder.priv_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_a").unwrap(),
-            [contains],
-        ))?;
+        let dp_out = partition::partition_with_ordering(&shape, &ordering)
+            .expect("DP must produce a partition on a valid topo ordering");
 
-        let b_out = builder.priv_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_b").unwrap(),
-            [a_out],
-        ))?;
-
-        let c_out = builder.priv_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_c").unwrap(),
-            [b_out],
-        ))?;
-
-        let _d_out = builder.pub_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_d").unwrap(),
-            [c_out],
-        ))?;
-
-        let solved = builder.solve()?;
-        let solution = solved.solution();
-
-        // Expected: exactly 3 PODs for a 5-statement chain with max_priv=2
-        // (5 statements / 2 per POD = 3 PODs)
-        assert_eq!(
-            solution.pod_count, 3,
-            "Expected exactly 3 PODs for 5-statement chain with max_priv=2"
+        eprintln!(
+            "MILP-derived ordering: DP returns K={}, sizes={:?}",
+            dp_out.pod_count,
+            dp_out
+                .pod_statements
+                .iter()
+                .map(|v| v.len())
+                .collect::<Vec<_>>()
         );
 
-        // All 5 statements should be assigned across the PODs
-        let all_statements: BTreeSet<usize> = solution
-            .pod_statements
-            .iter()
-            .flat_map(|s| s.iter().copied())
-            .collect();
+        let prod_out = partition::partition(&shape).expect("production partition");
+        eprintln!("production K = {}", prod_out.pod_count);
+
         assert_eq!(
-            all_statements,
-            (0..5).collect::<BTreeSet<_>>(),
-            "All 5 statements should be assigned"
+            dp_out.pod_count, 12,
+            "DP should reach the MILP optimum K=12 when given the MILP-derived ordering; \
+             if it doesn't, the gap is not purely an ordering-generation problem"
+        );
+    }
+
+    /// Cross-reference: where do heuristic's POD 3 statements (the
+    /// stmt-saturated, CPV-light cluster suspected to drive the K=13/K=12
+    /// gap) end up in MILP's K=12 partition? If MILP spreads them across
+    /// many PODs, the "import-heavy cluster needs to be split" hypothesis
+    /// holds; if MILP also keeps them together, the gap is elsewhere.
+    /// `#[ignore]`'d (calls SCIP).
+    #[test]
+    #[ignore]
+    fn cracked_refinery_pod3_dispersion() {
+        let shape: InputShape =
+            serde_json::from_str(CRACKED_REFINERY_FIXTURE).expect("fixture deserializes");
+        let n = shape.num_statements();
+
+        let heur = partition::partition(&shape).expect("heuristic partitions");
+        assert_eq!(heur.pod_count, 13);
+        let heur_pod3: Vec<usize> = heur.pod_statements[3].clone();
+        eprintln!("Heuristic POD 3: {} statements", heur_pod3.len());
+
+        let milp_out = partition_milp::solve_for_k(&shape, 12).expect("MILP must find K=12");
+        let mut milp_of: Vec<usize> = vec![usize::MAX; n];
+        for (p, stmts) in milp_out.pod_statements.iter().enumerate() {
+            for &s in stmts {
+                milp_of[s] = p;
+            }
+        }
+
+        let mut dispersion: std::collections::BTreeMap<usize, usize> =
+            std::collections::BTreeMap::new();
+        for &s in &heur_pod3 {
+            *dispersion.entry(milp_of[s]).or_insert(0) += 1;
+        }
+        eprintln!(
+            "  Their MILP POD distribution (milp_pod -> count): {:?}",
+            dispersion
         );
 
-        // Each POD should have at most max_priv_statements = 2
-        for (i, stmts) in solution.pod_statements.iter().enumerate() {
-            assert!(
-                stmts.len() <= 2,
-                "POD {} has {} statements, but max_priv=2: {:?}",
-                i,
-                stmts.len(),
-                stmts
+        // Also compute per-MILP-POD CPV totals to see how MILP balances.
+        eprintln!("\nMILP POD CPV totals:");
+        for (p, stmts) in milp_out.pod_statements.iter().enumerate() {
+            let cpv: usize = stmts
+                .iter()
+                .map(|&s| shape.costs[s].custom_pred_verifications)
+                .sum();
+            let n_stmts = stmts.len();
+            eprintln!("  POD {}: {} stmts, {} CPV", p, n_stmts, cpv);
+        }
+
+        // For comparison, heuristic CPV per POD.
+        eprintln!("\nHeuristic POD CPV totals:");
+        for (p, stmts) in heur.pod_statements.iter().enumerate() {
+            let cpv: usize = stmts
+                .iter()
+                .map(|&s| shape.costs[s].custom_pred_verifications)
+                .sum();
+            let n_stmts = stmts.len();
+            eprintln!("  POD {}: {} stmts, {} CPV", p, n_stmts, cpv);
+        }
+    }
+
+    /// Detailed partition-diff between heuristic K=13 and MILP K=12 for
+    /// cracked-refinery. Prints (1) the transfer matrix of statements,
+    /// (2) which heuristic POD is "dissolved" in MILP's solution, (3) for
+    /// each scattered statement, its heuristic ordering position and MILP
+    /// target POD, and (4) the heuristic-vs-MILP ordering positions to
+    /// gauge how far the heuristic ordering is from MILP's effective one.
+    /// `#[ignore]`'d (calls SCIP).
+    #[test]
+    #[ignore]
+    fn cracked_refinery_partition_diff() {
+        let shape: InputShape =
+            serde_json::from_str(CRACKED_REFINERY_FIXTURE).expect("fixture deserializes");
+        let n = shape.num_statements();
+
+        // Heuristic K=13 + its ordering. Run bin-packing directly so we can
+        // pin the ordering it was generated from for position-by-position
+        // comparison.
+        let identity: Vec<usize> = (0..n).collect();
+        let bp_ordering = partition::kahn_bin_packing(&shape, &identity)
+            .expect("bin-packing must produce an ordering");
+        let heur = partition::partition_with_ordering(&shape, &bp_ordering)
+            .expect("DP must produce a partition on bin-packing ordering");
+        assert_eq!(heur.pod_count, 13);
+        let mut heur_pod_of: Vec<usize> = vec![usize::MAX; n];
+        for (p, stmts) in heur.pod_statements.iter().enumerate() {
+            for &s in stmts {
+                heur_pod_of[s] = p;
+            }
+        }
+        let mut heur_pos: Vec<usize> = vec![usize::MAX; n];
+        for (i, &s) in bp_ordering.iter().enumerate() {
+            heur_pos[s] = i;
+        }
+
+        // MILP K=12 partition.
+        let milp = partition_milp::solve_for_k(&shape, 12).expect("MILP K=12 must be feasible");
+        let mut milp_pod_of: Vec<usize> = vec![usize::MAX; n];
+        for (p, stmts) in milp.pod_statements.iter().enumerate() {
+            for &s in stmts {
+                milp_pod_of[s] = p;
+            }
+        }
+
+        // Transfer matrix: rows = MILP POD, cols = heur POD.
+        let mut transfer = vec![vec![0_usize; 13]; 12];
+        for s in 0..n {
+            transfer[milp_pod_of[s]][heur_pod_of[s]] += 1;
+        }
+        eprintln!("Transfer matrix (rows=MILP POD, cols=heur POD):");
+        eprint!("            ");
+        for q in 0..13 {
+            eprint!("h{:<3} ", q);
+        }
+        eprintln!();
+        for (p, row) in transfer.iter().enumerate() {
+            eprint!("  MILP{:2}:  ", p);
+            for &v in row.iter() {
+                if v == 0 {
+                    eprint!("   . ");
+                } else {
+                    eprint!("{:4} ", v);
+                }
+            }
+            let total: usize = row.iter().sum();
+            eprintln!(" (total {})", total);
+        }
+
+        // Greedy max-overlap matching: for each MILP POD, find best
+        // heuristic POD. The unmatched heuristic POD is the "dissolved" one.
+        let mut milp_match = [usize::MAX; 12];
+        let mut used = [false; 13];
+        let mut pairs: Vec<(usize, usize, usize)> = Vec::new();
+        for (p, row) in transfer.iter().enumerate() {
+            for (q, &v) in row.iter().enumerate() {
+                pairs.push((v, p, q));
+            }
+        }
+        pairs.sort_by(|a, b| b.0.cmp(&a.0));
+        for (_, p, q) in pairs {
+            if milp_match[p] == usize::MAX && !used[q] {
+                milp_match[p] = q;
+                used[q] = true;
+            }
+        }
+        let dissolved: Vec<usize> = (0..13).filter(|&q| !used[q]).collect();
+        eprintln!("MILP POD -> best-matching heur POD:");
+        for p in 0..12 {
+            let q = milp_match[p];
+            let kept = transfer[p][q];
+            let total: usize = transfer[p].iter().sum();
+            eprintln!(
+                "  MILP{:2} -> heur{:2}  ({} of {} kept; {} migrated in)",
+                p,
+                q,
+                kept,
+                total,
+                total - kept
+            );
+        }
+        eprintln!("Dissolved heuristic PODs: {:?}", dissolved);
+
+        // Heuristic-POD-CPV vs MILP-POD-CPV (for context).
+        eprintln!("\nHeur POD CPV/stmt counts:");
+        for p in 0..13 {
+            let stmts = &heur.pod_statements[p];
+            let cpv: usize = stmts
+                .iter()
+                .map(|&s| shape.costs[s].custom_pred_verifications)
+                .sum();
+            eprintln!("  heur{:2}: {:2} stmts, {} CPV", p, stmts.len(), cpv);
+        }
+        eprintln!("MILP POD CPV/stmt counts:");
+        for p in 0..12 {
+            let stmts = &milp.pod_statements[p];
+            let cpv: usize = stmts
+                .iter()
+                .map(|&s| shape.costs[s].custom_pred_verifications)
+                .sum();
+            eprintln!("  MILP{:2}: {:2} stmts, {} CPV", p, stmts.len(), cpv);
+        }
+
+        // Construct a MILP-derived ordering: concatenate statements by
+        // (milp_pod_of, topo_pos). Use bp_ordering's position as the
+        // intra-POD topo order; it's a valid topo sort and stable.
+        let mut milp_ordering: Vec<usize> = (0..n).collect();
+        milp_ordering.sort_by_key(|&s| (milp_pod_of[s], heur_pos[s]));
+        let mut milp_pos: Vec<usize> = vec![usize::MAX; n];
+        for (i, &s) in milp_ordering.iter().enumerate() {
+            milp_pos[s] = i;
+        }
+
+        // Sanity check: is the MILP-derived ordering itself a valid topo
+        // ordering? (MILP's pod-precedence + bp_ordering tiebreak should
+        // give a valid one, but verify.)
+        let mut topo_valid = true;
+        for s in 0..n {
+            for dep in &shape.dep_edges[s] {
+                if let AbstractDep::Internal(d) = dep {
+                    if milp_pos[*d] >= milp_pos[s] {
+                        topo_valid = false;
+                        break;
+                    }
+                }
+            }
+            if !topo_valid {
+                break;
+            }
+        }
+        eprintln!("\nMILP-derived ordering is topo-valid: {}", topo_valid);
+
+        // DP K on the MILP-derived ordering.
+        if topo_valid {
+            let milp_dp = partition::partition_with_ordering(&shape, &milp_ordering);
+            eprintln!(
+                "DP K on MILP-derived ordering: {:?}",
+                milp_dp.as_ref().map(|o| o.pod_count)
             );
         }
 
-        // The output POD (last) must contain d_out(4) and it must be public
-        let output_pod_idx = solution.pod_count - 1;
-        assert!(
-            solution.pod_statements[output_pod_idx].contains(&4),
-            "Output POD should contain statement 4 (d_out), got {:?}",
-            solution.pod_statements[output_pod_idx]
-        );
-        assert!(
-            solution.pod_public_statements[output_pod_idx].contains(&4),
-            "Statement 4 (d_out) should be public in output POD"
+        // Position diff distribution: how far does each statement move
+        // between bp_ordering and milp_ordering?
+        let mut diffs: Vec<i64> = (0..n)
+            .map(|s| milp_pos[s] as i64 - heur_pos[s] as i64)
+            .collect();
+        let absdiffs: Vec<i64> = diffs.iter().map(|d| d.abs()).collect();
+        let max_diff = absdiffs.iter().copied().max().unwrap_or(0);
+        let mean_diff = absdiffs.iter().sum::<i64>() as f64 / n as f64;
+        let unchanged = diffs.iter().filter(|&&d| d == 0).count();
+        let small_move = diffs.iter().filter(|&&d| d.abs() <= 5 && d != 0).count();
+        let medium_move = diffs
+            .iter()
+            .filter(|&&d| d.abs() > 5 && d.abs() <= 50)
+            .count();
+        let large_move = diffs.iter().filter(|&&d| d.abs() > 50).count();
+        eprintln!(
+            "Position diff bp->milp: max={} mean={:.1} | unchanged={} small(<=5)={} medium(6-50)={} large(>50)={}",
+            max_diff, mean_diff, unchanged, small_move, medium_move, large_move
         );
 
-        // Prove and verify all PODs
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
+        // Kendall tau (number of pairwise inversions) is too expensive at
+        // n=345; report a sample-based proxy instead: how often does
+        // (bp_ordering[i] < bp_ordering[j]) match
+        // (milp_pos[bp_ordering[i]] < milp_pos[bp_ordering[j]])?
+        let sample_pairs = 5000;
+        let mut rng_state = 0xDEADBEEF_u64;
+        let mut concord = 0;
+        let mut discord = 0;
+        for _ in 0..sample_pairs {
+            rng_state = rng_state
+                .wrapping_mul(6364136223846793005)
+                .wrapping_add(1442695040888963407);
+            let i = (rng_state >> 33) as usize % n;
+            rng_state = rng_state
+                .wrapping_mul(6364136223846793005)
+                .wrapping_add(1442695040888963407);
+            let j = (rng_state >> 33) as usize % n;
+            if i == j {
+                continue;
+            }
+            let s_i = bp_ordering[i.min(j)];
+            let s_j = bp_ordering[i.max(j)];
+            if milp_pos[s_i] < milp_pos[s_j] {
+                concord += 1;
+            } else {
+                discord += 1;
+            }
+        }
+        let total = concord + discord;
+        eprintln!(
+            "Pairwise concord (random {} pairs): {} concordant / {} = {:.1}%",
+            total,
+            concord,
+            total,
+            100.0 * concord as f64 / total as f64
+        );
+        diffs.sort();
+        eprintln!(
+            "Diff percentiles: p10={} p25={} p50={} p75={} p90={}",
+            diffs[n * 10 / 100],
+            diffs[n * 25 / 100],
+            diffs[n * 50 / 100],
+            diffs[n * 75 / 100],
+            diffs[n * 90 / 100],
+        );
 
-        for (i, pod) in result.pods.iter().enumerate() {
-            pod.pod
-                .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
+        // Probe: how few statements need to be at their MILP position
+        // (with the rest preserving bp_ordering's relative order) for
+        // DP K to drop to 12? Largest movers first.
+        let mut indices: Vec<usize> = (0..n).collect();
+        indices
+            .sort_by_key(|&s| std::cmp::Reverse((milp_pos[s] as i64 - heur_pos[s] as i64).abs()));
+        eprintln!("\nGreedy 'minimum critical commits' probe:");
+        eprintln!("  Commit top-K largest movers to their MILP positions; rest fill in bp_ordering order.");
+        eprintln!("  {:<10} {:<10} {:<8}", "committed", "topo_ok", "dp_k");
+        for k_commit in [0, 5, 10, 20, 40, 60, 80, 120, 160, 200, 240, 345] {
+            let k_commit = k_commit.min(n);
+            let committed: HashSet<usize> = indices.iter().take(k_commit).copied().collect();
+            let mut slot: Vec<usize> = vec![usize::MAX; n];
+            for &s in &committed {
+                slot[milp_pos[s]] = s;
+            }
+            let remaining: Vec<usize> = bp_ordering
+                .iter()
+                .copied()
+                .filter(|s| !committed.contains(s))
+                .collect();
+            let mut ri = 0;
+            for s in slot.iter_mut() {
+                if *s == usize::MAX {
+                    *s = remaining[ri];
+                    ri += 1;
+                }
+            }
+            // Topo check.
+            let mut pos_new = vec![0_usize; n];
+            for (i, &s) in slot.iter().enumerate() {
+                pos_new[s] = i;
+            }
+            let mut topo_ok = true;
+            for s in 0..n {
+                for dep in &shape.dep_edges[s] {
+                    if let AbstractDep::Internal(d) = dep {
+                        if pos_new[*d] >= pos_new[s] {
+                            topo_ok = false;
+                            break;
+                        }
+                    }
+                }
+                if !topo_ok {
+                    break;
+                }
+            }
+            let dp_k = if topo_ok {
+                partition::partition_with_ordering(&shape, &slot).map(|o| o.pod_count)
+            } else {
+                None
+            };
+            eprintln!("  {:<10} {:<10} {:?}", k_commit, topo_ok, dp_k);
         }
 
-        Ok(())
+        // List statements that move and where they go.
+        let mut moves: Vec<(usize, usize, usize)> = Vec::new();
+        for s in 0..n {
+            let q = heur_pod_of[s];
+            let p = milp_pod_of[s];
+            let q_target_for_p = milp_match[p];
+            if q != q_target_for_p {
+                moves.push((s, q, p));
+            }
+        }
+        eprintln!(
+            "\nStatements moving across heur->MILP (relative to greedy POD match): {}",
+            moves.len()
+        );
+        eprintln!("(stmt | heur_pod heur_pos | milp_pod milp_match | cpv | mp | mst | sb | cps)");
+        for (s, heur_p, milp_p) in &moves {
+            let c = &shape.costs[*s];
+            eprintln!(
+                "  s={:3} | heur{:2} pos {:3} | MILP{:2} (match heur{:2}) | cpv={} mp_s={} mp_m={} mst_s={} mst_m={} sb={} cps={}",
+                s,
+                heur_p,
+                heur_pos[*s],
+                milp_p,
+                milp_match[*milp_p],
+                c.custom_pred_verifications,
+                c.merkle_proofs_small,
+                c.merkle_proofs_medium,
+                c.merkle_state_transitions_small,
+                c.merkle_state_transitions_medium,
+                c.signed_by,
+                c.custom_predicates_ids.len(),
+            );
+        }
     }
 
+    /// Probe: characterise cracked-refinery's CPV block structure. For
+    /// each CPV head, count direct Internal deps that have no other
+    /// consumers (the "exclusive body") and report the block-size
+    /// histogram. Then estimate the minimum K achievable by atomic
+    /// block packing (CPV head + body bundled inseparably) via
+    /// first-fit-decreasing. Used to decide whether a pure block-based
+    /// generator is viable for this fixture or whether atomic packing
+    /// would regress.
     #[test]
-    fn test_diamond_dependencies_across_pods() -> Result<()> {
-        // Verifies that diamond-shaped dependencies work across PODs.
-        //
-        // Diamond structure:
-        //           a_out (output)
-        //          /      \
-        //      b_out      c_out
-        //          \      /
-        //          contains
-        //
-        // Where a_out depends on BOTH b_out and c_out, creating a diamond.
-        // The solver may distribute statements across PODs in various ways,
-        // as long as dependencies are satisfied.
+    fn cracked_refinery_block_size_probe() {
+        let shape: InputShape =
+            serde_json::from_str(CRACKED_REFINERY_FIXTURE).expect("fixture deserializes");
+        let n = shape.num_statements();
+
+        // Inverse of dep_edges: for each statement, who depends on it.
+        let mut consumers: Vec<Vec<usize>> = vec![Vec::new(); n];
+        for (i, edges) in shape.dep_edges.iter().enumerate() {
+            for dep in edges {
+                if let AbstractDep::Internal(d) = dep {
+                    consumers[*d].push(i);
+                }
+            }
+        }
+
+        // For each CPV head, compute exclusive vs shared direct-body counts.
+        let mut block_sizes: Vec<usize> = Vec::new();
+        let mut cpv_count: usize = 0;
+        let mut total_exclusive: usize = 0;
+        let mut total_shared: usize = 0;
+        let mut total_external_deps: usize = 0;
+        for (i, cost) in shape.costs.iter().enumerate() {
+            if cost.custom_pred_verifications == 0 {
+                continue;
+            }
+            cpv_count += 1;
+            let mut excl = 0;
+            let mut shared = 0;
+            let mut ext = 0;
+            for dep in &shape.dep_edges[i] {
+                match dep {
+                    AbstractDep::Internal(d) => {
+                        if consumers[*d].len() == 1 && consumers[*d][0] == i {
+                            excl += 1;
+                        } else {
+                            shared += 1;
+                        }
+                    }
+                    AbstractDep::External { .. } => ext += 1,
+                }
+            }
+            block_sizes.push(1 + excl);
+            total_exclusive += excl;
+            total_shared += shared;
+            total_external_deps += ext;
+        }
+
+        eprintln!("Total CPV heads: {}", cpv_count);
+        eprintln!("Total exclusive direct body stmts: {}", total_exclusive);
+        eprintln!("Total shared direct body stmts: {}", total_shared);
+        eprintln!(
+            "Total external direct deps on CPVs: {}",
+            total_external_deps
+        );
+        let mean_block: f64 = block_sizes.iter().sum::<usize>() as f64 / cpv_count.max(1) as f64;
+        eprintln!("Mean block size (head + exclusive body): {:.2}", mean_block);
+
+        let mut hist: std::collections::BTreeMap<usize, usize> = std::collections::BTreeMap::new();
+        for &sz in &block_sizes {
+            *hist.entry(sz).or_insert(0) += 1;
+        }
+        eprintln!("Block-size histogram:");
+        for (sz, count) in &hist {
+            eprintln!("  size {:>2}: {:>3} blocks", sz, count);
+        }
+
+        // Estimate K under atomic block-based packing via FFD.
+        // Caps: max_statements stmts/POD, max_custom_pred_verifications
+        // CPVs/POD. Each block contributes block_size stmts and 1 CPV.
+        let p = &shape.params;
+        let max_stmts = p.max_statements;
+        let max_cpvs = p.max_custom_predicate_verifications;
+        let mut sorted = block_sizes.clone();
+        sorted.sort_unstable();
+        sorted.reverse();
+        let mut bins: Vec<(usize, usize)> = Vec::new();
+        for &b in &sorted {
+            let mut placed = false;
+            for bin in bins.iter_mut() {
+                if bin.0 + b <= max_stmts && bin.1 < max_cpvs {
+                    bin.0 += b;
+                    bin.1 += 1;
+                    placed = true;
+                    break;
+                }
+            }
+            if !placed {
+                bins.push((b, 1));
+            }
+        }
+        eprintln!("FFD atomic block packing (CPV blocks only, ignoring all other stmts):");
+        eprintln!("  bins used: {}", bins.len());
+        let cpv_floor = cpv_count.div_ceil(max_cpvs);
+        eprintln!(
+            "  CPV-cap floor (cpvs / cap): {} (current heuristic K = 13, MILP K = 12)",
+            cpv_floor
+        );
+
+        // Verdict.
+        let max_block = sorted.first().copied().unwrap_or(0);
+        let threshold_block = max_stmts / max_cpvs; // 40/8 = 5
+        let total_block_stmts: usize = sorted.iter().sum();
+        let free_stmts = n.saturating_sub(total_block_stmts);
+        eprintln!(
+            "Block stmts: {}; free (non-block) stmts: ~{}",
+            total_block_stmts, free_stmts,
+        );
+        eprintln!(
+            "Largest block = {}; uniform-block threshold for atomic safety = {}",
+            max_block, threshold_block,
+        );
+        if bins.len() <= cpv_floor {
+            eprintln!(
+                "  -> FFD on CPV blocks alone matches CPV floor (={}). Atomic block-based packing is",
+                cpv_floor,
+            );
+            eprintln!(
+                "     viable in principle for this fixture's actual block mix; the open question is"
+            );
+            eprintln!(
+                "     whether the {} free statements + topo + other-resource caps fit within {} PODs.",
+                free_stmts, bins.len(),
+            );
+            if max_block > threshold_block {
+                eprintln!(
+                    "  -> Caveat: largest block = {} > {} means a uniform workload at this block size"
+                , max_block, threshold_block);
+                eprintln!(
+                    "     would regress. The mix (esp. {} small blocks at sizes 1-2) is what makes it work.",
+                    hist.get(&1).copied().unwrap_or(0) + hist.get(&2).copied().unwrap_or(0),
+                );
+            }
+        } else {
+            eprintln!(
+                "  -> FFD on CPV blocks alone needs {} bins > CPV floor {}: atomic block-based",
+                bins.len(),
+                cpv_floor,
+            );
+            eprintln!("     packing regresses.");
+        }
+    }
+
+    /// Probe: does pre-emptively synth-ifying every external (input)
+    /// statement on cracked-refinery drop K from 13 to 12? The fixture
+    /// has 2 external pods x 1 input statement x 1 consumer each (so the
+    /// existing 2+-consumer and feasibility republish rules don't fire).
+    /// User's hypothesis is that proactively republishing into the chain
+    /// lets the partition place the externals in PODs with spare capacity
+    /// instead of forcing the consumer POD to open the external pod.
+    #[test]
+    fn cracked_refinery_preemptive_synth_probe() {
+        use cost::StatementCost;
+        let shape: InputShape =
+            serde_json::from_str(CRACKED_REFINERY_FIXTURE).expect("fixture deserializes");
+        let n_orig = shape.num_statements();
+        let num_statements = shape.statement_pod.len();
+        assert_eq!(
+            num_statements, 2,
+            "cracked-refinery has 2 external input statements"
+        );
+
+        // Build augmented dep_edges: replace each External with
+        // Internal(n_orig + statement).
+        let mut new_dep_edges: Vec<Vec<AbstractDep>> = shape
+            .dep_edges
+            .iter()
+            .map(|edges| {
+                edges
+                    .iter()
+                    .map(|dep| match dep {
+                        AbstractDep::Internal(d) => AbstractDep::Internal(*d),
+                        AbstractDep::External { statement, .. } => {
+                            AbstractDep::Internal(n_orig + *statement)
+                        }
+                    })
+                    .collect()
+            })
+            .collect();
+        // Append synth statements with the External dep moved to them.
+        for statement in 0..num_statements {
+            new_dep_edges.push(vec![AbstractDep::External {
+                pod: shape.statement_pod[statement],
+                statement,
+            }]);
+        }
+        let mut new_costs: Vec<StatementCost> = shape.costs.clone();
+        for _ in 0..num_statements {
+            new_costs.push(StatementCost::default());
+        }
+        let augmented = InputShape {
+            costs: new_costs,
+            dep_edges: new_dep_edges,
+            output_public_indices: shape.output_public_indices.clone(),
+            num_external_pods: shape.num_external_pods,
+            statement_pod: shape.statement_pod.clone(),
+            params: shape.params.clone(),
+        };
+
+        let baseline = partition::partition(&shape).expect("baseline partitions");
+        let augmented_out = partition::partition(&augmented).expect("augmented partitions");
+        eprintln!(
+            "baseline K = {}, sizes = {:?}",
+            baseline.pod_count,
+            baseline
+                .pod_statements
+                .iter()
+                .map(|v| v.len())
+                .collect::<Vec<_>>(),
+        );
+        eprintln!(
+            "augmented (default prio) K = {}, sizes = {:?}",
+            augmented_out.pod_count,
+            augmented_out
+                .pod_statements
+                .iter()
+                .map(|v| v.len())
+                .collect::<Vec<_>>(),
+        );
+
+        // Probe synth-first priority: give synths priority 0..S, originals
+        // priority S..n. This forces bin-packing to place synths in the
+        // earliest segment with capacity.
+        let n_aug = augmented.num_statements();
+        let mut prio: Vec<usize> = vec![0; n_aug];
+        for (rank, p) in (n_orig..n_aug).enumerate() {
+            prio[p] = rank;
+        }
+        for (rank, slot) in prio.iter_mut().take(n_orig).enumerate() {
+            *slot = num_statements + rank;
+        }
+        let ord = partition::kahn_bin_packing(&augmented, &prio)
+            .expect("synth-first bin-packing must produce an ordering");
+        let synth_first_out = partition::partition_with_ordering(&augmented, &ord)
+            .expect("DP must partition synth-first ordering");
+        eprintln!(
+            "augmented + synth-first prio bp K = {}, sizes = {:?}",
+            synth_first_out.pod_count,
+            synth_first_out
+                .pod_statements
+                .iter()
+                .map(|v| v.len())
+                .collect::<Vec<_>>(),
+        );
+
+        // Per-POD CPV totals (synth-first augmented).
+        let cpvs: Vec<usize> = synth_first_out
+            .pod_statements
+            .iter()
+            .map(|stmts| {
+                stmts
+                    .iter()
+                    .map(|&s| augmented.costs[s].custom_pred_verifications)
+                    .sum()
+            })
+            .collect();
+        eprintln!("  synth-first per-POD CPV: {:?}", cpvs);
+
+        // Also try other generators with synth-first priority.
+        if let Some(ord_prio) = partition::kahn_with_priority(&augmented, &prio) {
+            if let Some(out) = partition::partition_with_ordering(&augmented, &ord_prio) {
+                eprintln!("augmented + synth-first prio kahn K = {}", out.pod_count);
+            }
+        }
+    }
+
+    /// A non-statement-table cap (`max_signed_by`) forces a multi-POD
+    /// split even though every POD has plenty of statement-table slack.
+    /// Confirms the partitioner respects per-resource caps beyond
+    /// `max_statements`.
+    #[test]
+    fn signed_by_limit_forces_multi_pod_split() {
+        use crate::{
+            backends::plonky2::{primitives::ec::schnorr::SecretKey, signer::Signer},
+            frontend::SignedDictBuilder,
+        };
 
         let params = Params {
-            max_statements: 6,
-            max_public_statements: 3,
-            // max_priv_statements = 3
-            max_input_pods: 4,
-            max_input_pods_public_statements: 20,
-            max_custom_predicate_verifications: 10,
+            max_signed_by: 2,
             ..Params::default()
         };
         let vd_set = &*MOCK_VD_SET;
-
-        // pred_a takes TWO custom statement arguments (b_out and c_out)
-        // pred_b and pred_c each take a Contains
-        // Note: AND clauses are newline-separated, not comma-separated
-        let module = load_module(
-            r#"
-            pred_b(X) = AND(Contains(X, "k", 1))
-            pred_c(X) = AND(Contains(X, "k", 1))
-            pred_a(X, Y) = AND(
-                pred_b(X)
-                pred_c(Y)
-            )
-            "#,
-            "test",
-            &params,
-            &[],
-        )
-        .expect("load module");
-        let batch = &module.batch;
+        let prover = MockProver {};
 
         let mut builder = MultiPodBuilder::new(&params, vd_set);
+        for i in 0..4_i64 {
+            let mut signed_builder = SignedDictBuilder::new(&params);
+            signed_builder.insert("id", i);
+            let signer = Signer(SecretKey((i as u32 + 1).into()));
+            let signed_dict = signed_builder.sign(&signer).expect("sign dict");
+            if i == 3 {
+                builder
+                    .pub_op(FrontendOp::dict_signed_by(&signed_dict))
+                    .expect("pub signed_by");
+            } else {
+                builder
+                    .priv_op(FrontendOp::dict_signed_by(&signed_dict))
+                    .expect("priv signed_by");
+            }
+        }
 
-        // Base: single contains statement (shared by both branches conceptually,
-        // but we need separate ones for pred_b and pred_c due to predicate signatures)
-        let dict = dict!({"k" => 1});
-        let contains = builder.priv_op(FrontendOp::dict_contains(dict, "k", 1))?;
-
-        // Left branch: b_out depends on contains
-        let b_out = builder.priv_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_b").unwrap(),
-            [contains.clone()],
-        ))?;
-
-        // Right branch: c_out depends on contains
-        let c_out = builder.priv_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_c").unwrap(),
-            [contains],
-        ))?;
-
-        // Top: a_out depends on BOTH b_out and c_out
-        let _a_out = builder.pub_op(FrontendOp::custom(
-            batch.predicate_ref_by_name("pred_a").unwrap(),
-            [b_out, c_out],
-        ))?;
-
-        let solved = builder.solve()?;
-        let solution = solved.solution();
-
-        // With 4 statements and max_priv=3, we need at least 2 PODs (ceil(4/3) = 2)
+        let solved = builder.solve().expect("should solve");
         assert_eq!(
-            solution.pod_count, 2,
-            "Expected exactly 2 PODs for diamond with max_priv=3"
+            solved.solution().pod_count,
+            2,
+            "4 SignedBy ops with max_signed_by=2 require exactly 2 PODs"
         );
-
-        // The output POD (last) must contain statement 3 (a_out) and it must be public
-        let output_pod_idx = solution.pod_count - 1;
-        assert!(
-            solution.pod_statements[output_pod_idx].contains(&3),
-            "Output POD should contain statement 3 (a_out), got {:?}",
-            solution.pod_statements[output_pod_idx]
-        );
-        assert!(
-            solution.pod_public_statements[output_pod_idx].contains(&3),
-            "Statement 3 (a_out) should be public in output POD"
-        );
-
-        // All statements should be covered exactly once across all PODs
-        let all_statements: BTreeSet<usize> = solution
-            .pod_statements
-            .iter()
-            .flat_map(|s| s.iter().copied())
-            .collect();
-        assert_eq!(
-            all_statements,
-            [0, 1, 2, 3].into_iter().collect(),
-            "All statements should be assigned to exactly one POD"
-        );
-
-        // Prove and verify all PODs - this validates dependencies are satisfied
-        let prover = MockProver {};
-        let result = solved.prove(&prover)?;
-
+        let result = solved.prove(&prover).expect("prove should succeed");
+        assert_eq!(result.pods.len(), 2);
         for (i, pod) in result.pods.iter().enumerate() {
             pod.pod
                 .verify()
-                .unwrap_or_else(|_| panic!("POD {} verification failed", i));
+                .unwrap_or_else(|e| panic!("POD {} verification failed: {:?}", i, e));
+        }
+    }
+
+    /// Hand-construct a `DependencyGraph` so we can exercise the pre-pass
+    /// without rigging up a `MainPodBuilder` and prover.
+    mod prepass_tests {
+        use crate::{
+            frontend::{
+                multi_pod::{
+                    build_shape_and_index,
+                    deps::{DependencyGraph, ExternalDependency, StatementSource},
+                    shape::AbstractDep,
+                },
+                Operation as FrontendOp,
+            },
+            middleware::{
+                Hash, NativeOperation, OperationAux, OperationType, Params, RawValue, Statement,
+                Value, ValueRef,
+            },
+        };
+
+        fn noop_op() -> FrontendOp {
+            FrontendOp(
+                OperationType::Native(NativeOperation::None),
+                vec![],
+                OperationAux::None,
+            )
         }
 
-        Ok(())
+        fn ext_statement(pod_seed: i64, val: i64) -> ExternalDependency {
+            // A unique pod hash per seed and a literal-Equal statement per
+            // value. The statement contents are arbitrary as long as
+            // different input statements hash differently.
+            ExternalDependency {
+                pod_hash: Hash::from(RawValue::from(pod_seed)),
+                statement: Statement::Equal(
+                    ValueRef::Literal(Value::from(val)),
+                    ValueRef::Literal(Value::from(val)),
+                ),
+            }
+        }
+
+        #[test]
+        fn single_consumer_does_not_republish() {
+            let ext = ext_statement(1, 1);
+            let deps = DependencyGraph {
+                statement_deps: vec![vec![StatementSource::External(ext)]],
+            };
+            let operations = vec![noop_op()];
+            let (shape, _index) =
+                build_shape_and_index(&operations, &deps, &[], &Params::default());
+            assert_eq!(shape.num_statements(), 1);
+            assert!(matches!(
+                shape.dep_edges[0][0],
+                AbstractDep::External { .. }
+            ));
+        }
+
+        #[test]
+        fn two_consumers_trigger_republish() {
+            let ext = ext_statement(1, 1);
+            let deps = DependencyGraph {
+                statement_deps: vec![
+                    vec![StatementSource::External(ext.clone())],
+                    vec![StatementSource::External(ext)],
+                ],
+            };
+            let operations = vec![noop_op(), noop_op()];
+            let (shape, _index) =
+                build_shape_and_index(&operations, &deps, &[], &Params::default());
+
+            assert_eq!(shape.num_statements(), 3);
+            let synth_idx = 2;
+            assert_eq!(shape.dep_edges[0], vec![AbstractDep::Internal(synth_idx)]);
+            assert_eq!(shape.dep_edges[1], vec![AbstractDep::Internal(synth_idx)]);
+            assert!(matches!(
+                shape.dep_edges[synth_idx][0],
+                AbstractDep::External { .. }
+            ));
+        }
+
+        #[test]
+        fn opportunistic_bundling_republishes_single_consumer_siblings() {
+            // External pod 1 has input statement A (2 consumers) and input
+            // statement B (1 consumer). A triggers the basic 2+-consumer
+            // rule. Opportunistic bundling should pull B in too: the
+            // synth-hosting POD pays for pod 1 once regardless, so
+            // bundling B costs only one extra statement slot and frees
+            // B's consumer from referencing pod 1 directly.
+            let ext_a = ext_statement(1, 1);
+            let ext_b = ext_statement(1, 2);
+            let deps = DependencyGraph {
+                statement_deps: vec![
+                    vec![StatementSource::External(ext_a.clone())],
+                    vec![StatementSource::External(ext_a)],
+                    vec![StatementSource::External(ext_b.clone())],
+                ],
+            };
+            let operations = vec![noop_op(), noop_op(), noop_op()];
+            let (shape, _index) =
+                build_shape_and_index(&operations, &deps, &[], &Params::default());
+
+            // 3 originals + 2 synths (one per input statement of pod 1) = 5.
+            assert_eq!(shape.num_statements(), 5);
+            // Every original's external dep should be rewritten to Internal.
+            for orig in 0..3 {
+                assert!(
+                    matches!(shape.dep_edges[orig][0], AbstractDep::Internal(_)),
+                    "orig {} should reference a synth after bundling",
+                    orig
+                );
+            }
+        }
+
+        #[test]
+        fn bundling_does_not_cross_pods() {
+            // Two separate external pods. Pod 1 has an input statement
+            // with 2 consumers (triggers republish). Pod 2 has an input
+            // statement with 1 consumer. Pod 2's input statement should
+            // NOT be republished; bundling is per-pod.
+            let p1 = ext_statement(1, 1);
+            let p2 = ext_statement(2, 1);
+            let deps = DependencyGraph {
+                statement_deps: vec![
+                    vec![StatementSource::External(p1.clone())],
+                    vec![StatementSource::External(p1)],
+                    vec![StatementSource::External(p2)],
+                ],
+            };
+            let operations = vec![noop_op(), noop_op(), noop_op()];
+            let (shape, _index) =
+                build_shape_and_index(&operations, &deps, &[], &Params::default());
+
+            // 3 originals + 1 synth (only pod 1's input statement) = 4.
+            assert_eq!(shape.num_statements(), 4);
+            // The pod-2 consumer keeps its External dep.
+            assert!(matches!(
+                shape.dep_edges[2][0],
+                AbstractDep::External { .. }
+            ));
+        }
+
+        #[test]
+        fn distinct_statements_get_distinct_synthetics() {
+            let ext_a = ext_statement(1, 1);
+            let ext_b = ext_statement(1, 2);
+            let deps = DependencyGraph {
+                statement_deps: vec![
+                    vec![StatementSource::External(ext_a.clone())],
+                    vec![StatementSource::External(ext_a)],
+                    vec![StatementSource::External(ext_b.clone())],
+                    vec![StatementSource::External(ext_b)],
+                ],
+            };
+            let operations = vec![noop_op(), noop_op(), noop_op(), noop_op()];
+            let (shape, index) = build_shape_and_index(&operations, &deps, &[], &Params::default());
+
+            assert_eq!(shape.num_statements(), 6);
+            assert_eq!(index.statements.len(), 2);
+            assert_eq!(shape.num_external_pods, 1);
+        }
     }
 }
