+++ src/frontend/multi_pod/tests/fixtures/cracked_refinery.json
